@@ -20,7 +20,7 @@
         }
     ],
     "require": {
-        "php": ">=5.5.9"
+        "php": "^7.2.5"
     },
     "autoload": {
         "psr-4": { "Symfony\\Component\\CssSelector\\": "" },
@@ -31,7 +31,7 @@
     "minimum-stability": "dev",
     "extra": {
         "branch-alias": {
-            "dev-master": "3.3-dev"
+            "dev-master": "5.1-dev"
         }
     }
 }
