@@ -6,51 +6,49 @@ namespace voku\helper;
 
 use Symfony\Component\CssSelector\CssSelectorConverter;
 
-/**
- * Class SelectorConverter
- *
- * @package voku\helper
- */
 class SelectorConverter
 {
-  protected static $compiled = array();
+    /**
+     * @var array
+     */
+    protected static $compiled = [];
 
-  /**
-   * @param string $selector
-   *
-   * @return mixed|string
-   *
-   * @throws \RuntimeException
-   */
-  public static function toXPath(string $selector)
-  {
-    if (isset(self::$compiled[$selector])) {
-      return self::$compiled[$selector];
+    /**
+     * @param string $selector
+     *
+     * @throws \RuntimeException
+     *
+     * @return mixed|string
+     */
+    public static function toXPath(string $selector)
+    {
+        if (isset(self::$compiled[$selector])) {
+            return self::$compiled[$selector];
+        }
+
+        // Select DOMText
+        if ($selector === 'text') {
+            return '//text()';
+        }
+
+        // Select DOMComment
+        if ($selector === 'comment') {
+            return '//comment()';
+        }
+
+        if (\strpos($selector, '//') === 0) {
+            return $selector;
+        }
+
+        if (!\class_exists(CssSelectorConverter::class)) {
+            throw new \RuntimeException('Unable to filter with a CSS selector as the Symfony CssSelector 2.8+ is not installed (you can use filterXPath instead).');
+        }
+
+        $converter = new CssSelectorConverter(true);
+
+        $xPathQuery = $converter->toXPath($selector);
+        self::$compiled[$selector] = $xPathQuery;
+
+        return $xPathQuery;
     }
-
-    // Select DOMText
-    if ($selector === 'text') {
-      return '//text()';
-    }
-
-    // Select DOMComment
-    if ($selector === 'comment') {
-      return '//comment()';
-    }
-
-    if (\strpos($selector, '//') === 0) {
-      return $selector;
-    }
-
-    if (!\class_exists(CssSelectorConverter::class)) {
-      throw new \RuntimeException('Unable to filter with a CSS selector as the Symfony CssSelector 2.8+ is not installed (you can use filterXPath instead).');
-    }
-
-    $converter = new CssSelectorConverter(true);
-
-    $xPathQuery = $converter->toXPath($selector);
-    self::$compiled[$selector] = $xPathQuery;
-
-    return $xPathQuery;
-  }
 }
