@@ -24,33 +24,25 @@ use Symfony\Component\CssSelector\Parser\Token;
 class SyntaxErrorException extends ParseException
 {
     /**
-     * @param string $expectedValue
-     * @param Token  $foundToken
-     *
      * @return self
      */
-    public static function unexpectedToken($expectedValue, Token $foundToken)
+    public static function unexpectedToken(string $expectedValue, Token $foundToken)
     {
         return new self(sprintf('Expected %s, but %s found.', $expectedValue, $foundToken));
     }
 
     /**
-     * @param string $pseudoElement
-     * @param string $unexpectedLocation
-     *
      * @return self
      */
-    public static function pseudoElementFound($pseudoElement, $unexpectedLocation)
+    public static function pseudoElementFound(string $pseudoElement, string $unexpectedLocation)
     {
         return new self(sprintf('Unexpected pseudo-element "::%s" found %s.', $pseudoElement, $unexpectedLocation));
     }
 
     /**
-     * @param int $position
-     *
      * @return self
      */
-    public static function unclosedString($position)
+    public static function unclosedString(int $position)
     {
         return new self(sprintf('Unclosed/invalid string at %s.', $position));
     }
