@@ -29,51 +29,23 @@ class Specificity
     const B_FACTOR = 10;
     const C_FACTOR = 1;
 
-    /**
-     * @var int
-     */
     private $a;
-
-    /**
-     * @var int
-     */
     private $b;
-
-    /**
-     * @var int
-     */
     private $c;
 
-    /**
-     * Constructor.
-     *
-     * @param int $a
-     * @param int $b
-     * @param int $c
-     */
-    public function __construct($a, $b, $c)
+    public function __construct(int $a, int $b, int $c)
     {
         $this->a = $a;
         $this->b = $b;
         $this->c = $c;
     }
 
-    /**
-     * @param Specificity $specificity
-     *
-     * @return self
-     */
-    public function plus(Specificity $specificity)
+    public function plus(self $specificity): self
     {
         return new self($this->a + $specificity->a, $this->b + $specificity->b, $this->c + $specificity->c);
     }
 
-    /**
-     * Returns global specificity value.
-     *
-     * @return int
-     */
-    public function getValue()
+    public function getValue(): int
     {
         return $this->a * self::A_FACTOR + $this->b * self::B_FACTOR + $this->c * self::C_FACTOR;
     }
@@ -81,12 +53,8 @@ class Specificity
     /**
      * Returns -1 if the object specificity is lower than the argument,
      * 0 if they are equal, and 1 if the argument is lower.
-     *
-     * @param Specificity $specificity
-     *
-     * @return int
      */
-    public function compareTo(Specificity $specificity)
+    public function compareTo(self $specificity): int
     {
         if ($this->a !== $specificity->a) {
             return $this->a > $specificity->a ? 1 : -1;
