@@ -26,11 +26,5 @@ use Symfony\Component\CssSelector\Parser\TokenStream;
  */
 interface HandlerInterface
 {
-    /**
-     * @param Reader      $reader
-     * @param TokenStream $stream
-     *
-     * @return bool
-     */
-    public function handle(Reader $reader, TokenStream $stream);
+    public function handle(Reader $reader, TokenStream $stream): bool;
 }
