@@ -31,7 +31,7 @@ class ClassParser implements ParserInterface
     /**
      * {@inheritdoc}
      */
-    public function parse($source)
+    public function parse(string $source): array
     {
         // Matches an optional namespace, optional element, and required class
         // $source = 'test|input.ab6bd_field';
@@ -41,11 +41,11 @@ class ClassParser implements ParserInterface
         //     2 => string 'input' (length=5)
         //     3 => string 'ab6bd_field' (length=11)
         if (preg_match('/^(?:([a-z]++)\|)?+([\w-]++|\*)?+\.([\w-]++)$/i', trim($source), $matches)) {
-            return array(
+            return [
                 new SelectorNode(new ClassNode(new ElementNode($matches[1] ?: null, $matches[2] ?: null), $matches[3])),
-            );
+            ];
         }
 
-        return array();
+        return [];
     }
 }
