@@ -17,1412 +17,1569 @@ namespace voku\helper;
  *
  * Ideas:
  * - http://perfectionkills.com/optimizing-html/
- *
- * @package voku\helper
  */
-class HtmlMin
+class HtmlMin implements HtmlMinInterface
 {
-  /**
-   * @var string
-   */
-  private static $regExSpace = "/[[:space:]]{2,}|[\r\n]+/u";
-
-  /**
-   * @var array
-   */
-  private static $optional_end_tags = [
-      'html',
-      'head',
-      'body',
-  ];
-
-  /**
-   * // https://mathiasbynens.be/demo/javascript-mime-type
-   * // https://developer.mozilla.org/en/docs/Web/HTML/Element/script#attr-type
-   *
-   * @var array
-   */
-  private static $executableScriptsMimeTypes = [
-      'text/javascript'          => '',
-      'text/ecmascript'          => '',
-      'text/jscript'             => '',
-      'application/javascript'   => '',
-      'application/x-javascript' => '',
-      'application/ecmascript'   => '',
-  ];
-
-  private static $selfClosingTags = [
-      'area',
-      'base',
-      'basefont',
-      'br',
-      'col',
-      'command',
-      'embed',
-      'frame',
-      'hr',
-      'img',
-      'input',
-      'isindex',
-      'keygen',
-      'link',
-      'meta',
-      'param',
-      'source',
-      'track',
-      'wbr',
-  ];
-
-  private static $trimWhitespaceFromTags = [
-      'article' => '',
-      'br'      => '',
-      'div'     => '',
-      'footer'  => '',
-      'hr'      => '',
-      'nav'     => '',
-      'p'       => '',
-      'script'  => '',
-  ];
-
-  /**
-   * @var array
-   */
-  private static $booleanAttributes = [
-      'allowfullscreen' => '',
-      'async'           => '',
-      'autofocus'       => '',
-      'autoplay'        => '',
-      'checked'         => '',
-      'compact'         => '',
-      'controls'        => '',
-      'declare'         => '',
-      'default'         => '',
-      'defaultchecked'  => '',
-      'defaultmuted'    => '',
-      'defaultselected' => '',
-      'defer'           => '',
-      'disabled'        => '',
-      'enabled'         => '',
-      'formnovalidate'  => '',
-      'hidden'          => '',
-      'indeterminate'   => '',
-      'inert'           => '',
-      'ismap'           => '',
-      'itemscope'       => '',
-      'loop'            => '',
-      'multiple'        => '',
-      'muted'           => '',
-      'nohref'          => '',
-      'noresize'        => '',
-      'noshade'         => '',
-      'novalidate'      => '',
-      'nowrap'          => '',
-      'open'            => '',
-      'pauseonexit'     => '',
-      'readonly'        => '',
-      'required'        => '',
-      'reversed'        => '',
-      'scoped'          => '',
-      'seamless'        => '',
-      'selected'        => '',
-      'sortable'        => '',
-      'truespeed'       => '',
-      'typemustmatch'   => '',
-      'visible'         => '',
-  ];
-  /**
-   * @var array
-   */
-  private static $skipTagsForRemoveWhitespace = [
-      'code',
-      'pre',
-      'script',
-      'style',
-      'textarea',
-  ];
-
-  /**
-   * @var array
-   */
-  private $protectedChildNodes = [];
-
-  /**
-   * @var string
-   */
-  private $protectedChildNodesHelper = 'html-min--voku--saved-content';
-
-  /**
-   * @var bool
-   */
-  private $doOptimizeViaHtmlDomParser = true;
-
-  /**
-   * @var bool
-   */
-  private $doOptimizeAttributes = true;
-
-  /**
-   * @var bool
-   */
-  private $doRemoveComments = true;
-
-  /**
-   * @var bool
-   */
-  private $doRemoveWhitespaceAroundTags = false;
-
-  /**
-   * @var bool
-   */
-  private $doRemoveOmittedQuotes = true;
-
-  /**
-   * @var bool
-   */
-  private $doRemoveOmittedHtmlTags = true;
-
-  /**
-   * @var bool
-   */
-  private $doRemoveHttpPrefixFromAttributes = false;
-
-  /**
-   * @var array
-   */
-  private $domainsToRemoveHttpPrefixFromAttributes = [
-      'google.com',
-      'google.de',
-  ];
-
-  /**
-   * @var bool
-   */
-  private $doSortCssClassNames = true;
-
-  /**
-   * @var bool
-   */
-  private $doSortHtmlAttributes = true;
-
-  /**
-   * @var bool
-   */
-  private $doRemoveDeprecatedScriptCharsetAttribute = true;
-
-  /**
-   * @var bool
-   */
-  private $doRemoveDefaultAttributes = false;
-
-  /**
-   * @var bool
-   */
-  private $doRemoveDeprecatedAnchorName = true;
-
-  /**
-   * @var bool
-   */
-  private $doRemoveDeprecatedTypeFromStylesheetLink = true;
-
-  /**
-   * @var bool
-   */
-  private $doRemoveDeprecatedTypeFromScriptTag = true;
-
-  /**
-   * @var bool
-   */
-  private $doRemoveValueFromEmptyInput = true;
-
-  /**
-   * @var bool
-   */
-  private $doRemoveEmptyAttributes = true;
-
-  /**
-   * @var bool
-   */
-  private $doSumUpWhitespace = true;
-
-  /**
-   * @var bool
-   */
-  private $doRemoveSpacesBetweenTags = false;
-
-  /**
-   * @var
-   */
-  private $withDocType;
-
-  /**
-   * HtmlMin constructor.
-   */
-  public function __construct()
-  {
-  }
-
-  /**
-   * @param boolean $doOptimizeAttributes
-   *
-   * @return $this
-   */
-  public function doOptimizeAttributes(bool $doOptimizeAttributes = true)
-  {
-    $this->doOptimizeAttributes = $doOptimizeAttributes;
-
-    return $this;
-  }
-
-  /**
-   * @param boolean $doOptimizeViaHtmlDomParser
-   *
-   * @return $this
-   */
-  public function doOptimizeViaHtmlDomParser(bool $doOptimizeViaHtmlDomParser = true)
-  {
-    $this->doOptimizeViaHtmlDomParser = $doOptimizeViaHtmlDomParser;
-
-    return $this;
-  }
-
-  /**
-   * @param boolean $doRemoveComments
-   *
-   * @return $this
-   */
-  public function doRemoveComments(bool $doRemoveComments = true)
-  {
-    $this->doRemoveComments = $doRemoveComments;
-
-    return $this;
-  }
-
-  /**
-   * @param boolean $doRemoveDefaultAttributes
-   *
-   * @return $this
-   */
-  public function doRemoveDefaultAttributes(bool $doRemoveDefaultAttributes = true)
-  {
-    $this->doRemoveDefaultAttributes = $doRemoveDefaultAttributes;
-
-    return $this;
-  }
-
-  /**
-   * @param boolean $doRemoveDeprecatedAnchorName
-   *
-   * @return $this
-   */
-  public function doRemoveDeprecatedAnchorName(bool $doRemoveDeprecatedAnchorName = true)
-  {
-    $this->doRemoveDeprecatedAnchorName = $doRemoveDeprecatedAnchorName;
-
-    return $this;
-  }
-
-  /**
-   * @param boolean $doRemoveDeprecatedScriptCharsetAttribute
-   *
-   * @return $this
-   */
-  public function doRemoveDeprecatedScriptCharsetAttribute(bool $doRemoveDeprecatedScriptCharsetAttribute = true)
-  {
-    $this->doRemoveDeprecatedScriptCharsetAttribute = $doRemoveDeprecatedScriptCharsetAttribute;
-
-    return $this;
-  }
-
-  /**
-   * @param boolean $doRemoveDeprecatedTypeFromScriptTag
-   *
-   * @return $this
-   */
-  public function doRemoveDeprecatedTypeFromScriptTag(bool $doRemoveDeprecatedTypeFromScriptTag = true)
-  {
-    $this->doRemoveDeprecatedTypeFromScriptTag = $doRemoveDeprecatedTypeFromScriptTag;
-
-    return $this;
-  }
-
-  /**
-   * @param boolean $doRemoveDeprecatedTypeFromStylesheetLink
-   *
-   * @return $this
-   */
-  public function doRemoveDeprecatedTypeFromStylesheetLink(bool $doRemoveDeprecatedTypeFromStylesheetLink = true)
-  {
-    $this->doRemoveDeprecatedTypeFromStylesheetLink = $doRemoveDeprecatedTypeFromStylesheetLink;
-
-    return $this;
-  }
-
-  /**
-   * @param boolean $doRemoveEmptyAttributes
-   *
-   * @return $this
-   */
-  public function doRemoveEmptyAttributes(bool $doRemoveEmptyAttributes = true)
-  {
-    $this->doRemoveEmptyAttributes = $doRemoveEmptyAttributes;
-
-    return $this;
-  }
-
-  /**
-   * @param boolean $doRemoveHttpPrefixFromAttributes
-   *
-   * @return $this
-   */
-  public function doRemoveHttpPrefixFromAttributes(bool $doRemoveHttpPrefixFromAttributes = true)
-  {
-    $this->doRemoveHttpPrefixFromAttributes = $doRemoveHttpPrefixFromAttributes;
-
-    return $this;
-  }
-
-  /**
-   * @param boolean $doRemoveSpacesBetweenTags
-   *
-   * @return $this
-   */
-  public function doRemoveSpacesBetweenTags(bool $doRemoveSpacesBetweenTags = true)
-  {
-    $this->doRemoveSpacesBetweenTags = $doRemoveSpacesBetweenTags;
-
-    return $this;
-  }
-
-  /**
-   * @param boolean $doRemoveValueFromEmptyInput
-   *
-   * @return $this
-   */
-  public function doRemoveValueFromEmptyInput(bool $doRemoveValueFromEmptyInput = true)
-  {
-    $this->doRemoveValueFromEmptyInput = $doRemoveValueFromEmptyInput;
-
-    return $this;
-  }
-
-  /**
-   * @param boolean $doRemoveWhitespaceAroundTags
-   *
-   * @return $this
-   */
-  public function doRemoveWhitespaceAroundTags(bool $doRemoveWhitespaceAroundTags = true)
-  {
-    $this->doRemoveWhitespaceAroundTags = $doRemoveWhitespaceAroundTags;
-
-    return $this;
-  }
-
-  /**
-   * @param bool $doRemoveOmittedQuotes
-   *
-   * @return $this
-   */
-  public function doRemoveOmittedQuotes(bool $doRemoveOmittedQuotes = true)
-  {
-    $this->doRemoveOmittedQuotes = $doRemoveOmittedQuotes;
-
-    return $this;
-  }
-
-  /**
-   * @param bool $doRemoveOmittedHtmlTags
-   *
-   * @return $this
-   */
-  public function doRemoveOmittedHtmlTags(bool $doRemoveOmittedHtmlTags = true)
-  {
-    $this->doRemoveOmittedHtmlTags = $doRemoveOmittedHtmlTags;
-
-    return $this;
-  }
-
-  /**
-   * @param boolean $doSortCssClassNames
-   *
-   * @return $this
-   */
-  public function doSortCssClassNames(bool $doSortCssClassNames = true)
-  {
-    $this->doSortCssClassNames = $doSortCssClassNames;
-
-    return $this;
-  }
-
-  /**
-   * @param boolean $doSortHtmlAttributes
-   *
-   * @return $this
-   */
-  public function doSortHtmlAttributes(bool $doSortHtmlAttributes = true)
-  {
-    $this->doSortHtmlAttributes = $doSortHtmlAttributes;
-
-    return $this;
-  }
-
-  /**
-   * @param boolean $doSumUpWhitespace
-   *
-   * @return $this
-   */
-  public function doSumUpWhitespace(bool $doSumUpWhitespace = true)
-  {
-    $this->doSumUpWhitespace = $doSumUpWhitespace;
-
-    return $this;
-  }
-
-  private function domNodeAttributesToString(\DOMNode $node): string
-  {
-    # Remove quotes around attribute values, when allowed (<p class="foo"> → <p class=foo>)
-    $attrstr = '';
-    if ($node->attributes != null) {
-      foreach ($node->attributes as $attribute) {
-        $attrstr .= $attribute->name;
-
-        if (
-            $this->doOptimizeAttributes === true
-            &&
-            isset(self::$booleanAttributes[$attribute->name])
-        ) {
-          $attrstr .= ' ';
-          continue;
-        }
-
-        $attrstr .= '=';
-
-        # http://www.whatwg.org/specs/web-apps/current-work/multipage/syntax.html#attributes-0
-        $omitquotes = $this->doRemoveOmittedQuotes
-                      &&
-                      $attribute->value != ''
-                      &&
-                      0 == \preg_match('/["\'=<>` \t\r\n\f]+/', $attribute->value);
-
-        $attr_val = $attribute->value;
-        $attrstr .= ($omitquotes ? '' : '"') . $attr_val . ($omitquotes ? '' : '"');
-        $attrstr .= ' ';
-      }
+    /**
+     * @var string
+     */
+    private static $regExSpace = "/[[:space:]]{2,}|[\r\n]/u";
+
+    /**
+     * @var array
+     */
+    private static $optional_end_tags = [
+        'html',
+        'head',
+        'body',
+    ];
+
+    private static $selfClosingTags = [
+        'area',
+        'base',
+        'basefont',
+        'br',
+        'col',
+        'command',
+        'embed',
+        'frame',
+        'hr',
+        'img',
+        'input',
+        'isindex',
+        'keygen',
+        'link',
+        'meta',
+        'param',
+        'source',
+        'track',
+        'wbr',
+    ];
+
+    private static $trimWhitespaceFromTags = [
+        'article' => '',
+        'br'      => '',
+        'div'     => '',
+        'footer'  => '',
+        'hr'      => '',
+        'nav'     => '',
+        'p'       => '',
+        'script'  => '',
+    ];
+
+    /**
+     * @var array
+     */
+    private static $booleanAttributes = [
+        'allowfullscreen' => '',
+        'async'           => '',
+        'autofocus'       => '',
+        'autoplay'        => '',
+        'checked'         => '',
+        'compact'         => '',
+        'controls'        => '',
+        'declare'         => '',
+        'default'         => '',
+        'defaultchecked'  => '',
+        'defaultmuted'    => '',
+        'defaultselected' => '',
+        'defer'           => '',
+        'disabled'        => '',
+        'enabled'         => '',
+        'formnovalidate'  => '',
+        'hidden'          => '',
+        'indeterminate'   => '',
+        'inert'           => '',
+        'ismap'           => '',
+        'itemscope'       => '',
+        'loop'            => '',
+        'multiple'        => '',
+        'muted'           => '',
+        'nohref'          => '',
+        'noresize'        => '',
+        'noshade'         => '',
+        'novalidate'      => '',
+        'nowrap'          => '',
+        'open'            => '',
+        'pauseonexit'     => '',
+        'readonly'        => '',
+        'required'        => '',
+        'reversed'        => '',
+        'scoped'          => '',
+        'seamless'        => '',
+        'selected'        => '',
+        'sortable'        => '',
+        'truespeed'       => '',
+        'typemustmatch'   => '',
+        'visible'         => '',
+    ];
+
+    /**
+     * @var array
+     */
+    private static $skipTagsForRemoveWhitespace = [
+        'code',
+        'pre',
+        'script',
+        'style',
+        'textarea',
+    ];
+
+    /**
+     * @var array
+     */
+    private $protectedChildNodes = [];
+
+    /**
+     * @var string
+     */
+    private $protectedChildNodesHelper = 'html-min--voku--saved-content';
+
+    /**
+     * @var bool
+     */
+    private $doOptimizeViaHtmlDomParser = true;
+
+    /**
+     * @var bool
+     */
+    private $doOptimizeAttributes = true;
+
+    /**
+     * @var bool
+     */
+    private $doRemoveComments = true;
+
+    /**
+     * @var bool
+     */
+    private $doRemoveWhitespaceAroundTags = false;
+
+    /**
+     * @var bool
+     */
+    private $doRemoveOmittedQuotes = true;
+
+    /**
+     * @var bool
+     */
+    private $doRemoveOmittedHtmlTags = true;
+
+    /**
+     * @var bool
+     */
+    private $doRemoveHttpPrefixFromAttributes = false;
+
+    /**
+     * @var array
+     */
+    private $domainsToRemoveHttpPrefixFromAttributes = [
+        'google.com',
+        'google.de',
+    ];
+
+    /**
+     * @var bool
+     */
+    private $doSortCssClassNames = true;
+
+    /**
+     * @var bool
+     */
+    private $doSortHtmlAttributes = true;
+
+    /**
+     * @var bool
+     */
+    private $doRemoveDeprecatedScriptCharsetAttribute = true;
+
+    /**
+     * @var bool
+     */
+    private $doRemoveDefaultAttributes = false;
+
+    /**
+     * @var bool
+     */
+    private $doRemoveDeprecatedAnchorName = true;
+
+    /**
+     * @var bool
+     */
+    private $doRemoveDeprecatedTypeFromStylesheetLink = true;
+
+    /**
+     * @var bool
+     */
+    private $doRemoveDeprecatedTypeFromScriptTag = true;
+
+    /**
+     * @var bool
+     */
+    private $doRemoveValueFromEmptyInput = true;
+
+    /**
+     * @var bool
+     */
+    private $doRemoveEmptyAttributes = true;
+
+    /**
+     * @var bool
+     */
+    private $doSumUpWhitespace = true;
+
+    /**
+     * @var bool
+     */
+    private $doRemoveSpacesBetweenTags = false;
+
+    /**
+     * @var bool
+     */
+    private $keepBrokenHtml = false;
+
+    /**
+     * @var bool
+     */
+    private $withDocType = false;
+
+    /**
+     * @var HtmlMinDomObserverInterface[]|\SplObjectStorage
+     */
+    private $domLoopObservers;
+
+    /**
+     * @var int
+     */
+    private $protected_tags_counter = 0;
+
+    /**
+     * HtmlMin constructor.
+     */
+    public function __construct()
+    {
+        $this->domLoopObservers = new \SplObjectStorage();
+
+        $this->attachObserverToTheDomLoop(new HtmlMinDomObserverOptimizeAttributes());
     }
 
-    return \trim($attrstr);
-  }
+    /**
+     * @param HtmlMinDomObserverInterface $observer
+     *
+     * @return void
+     */
+    public function attachObserverToTheDomLoop(HtmlMinDomObserverInterface $observer)
+    {
+        $this->domLoopObservers->attach($observer);
+    }
 
-  /**
-   * @param \DOMNode $node
-   *
-   * @return bool
-   */
-  private function domNodeClosingTagOptional(\DOMNode $node): bool
-  {
-    $tag_name = $node->nodeName;
-    $nextSibling = $this->getNextSiblingOfTypeDOMElement($node);
+    /**
+     * @param bool $doOptimizeAttributes
+     *
+     * @return $this
+     */
+    public function doOptimizeAttributes(bool $doOptimizeAttributes = true): self
+    {
+        $this->doOptimizeAttributes = $doOptimizeAttributes;
 
-    // https://html.spec.whatwg.org/multipage/syntax.html#syntax-tag-omission
+        return $this;
+    }
 
-    // Implemented:
-    //
-    // A <p> element's end tag may be omitted if the p element is immediately followed by an address, article, aside, blockquote, details, div, dl, fieldset, figcaption, figure, footer, form, h1, h2, h3, h4, h5, h6, header, hgroup, hr, main, menu, nav, ol, p, pre, section, table, or ul element, or if there is no more content in the parent element and the parent element is an HTML element that is not an a, audio, del, ins, map, noscript, or video element, or an autonomous custom element.
-    // An <li> element's end tag may be omitted if the li element is immediately followed by another li element or if there is no more content in the parent element.
-    // A <td> element's end tag may be omitted if the td element is immediately followed by a td or th element, or if there is no more content in the parent element.
-    // An <option> element's end tag may be omitted if the option element is immediately followed by another option element, or if it is immediately followed by an optgroup element, or if there is no more content in the parent element.
-    // A <tr> element's end tag may be omitted if the tr element is immediately followed by another tr element, or if there is no more content in the parent element.
-    // A <th> element's end tag may be omitted if the th element is immediately followed by a td or th element, or if there is no more content in the parent element.
-    // A <dt> element's end tag may be omitted if the dt element is immediately followed by another dt element or a dd element.
-    // A <dd> element's end tag may be omitted if the dd element is immediately followed by another dd element or a dt element, or if there is no more content in the parent element.
-    // An <rp> element's end tag may be omitted if the rp element is immediately followed by an rt or rp element, or if there is no more content in the parent element.
+    /**
+     * @param bool $doOptimizeViaHtmlDomParser
+     *
+     * @return $this
+     */
+    public function doOptimizeViaHtmlDomParser(bool $doOptimizeViaHtmlDomParser = true): self
+    {
+        $this->doOptimizeViaHtmlDomParser = $doOptimizeViaHtmlDomParser;
 
-    // TODO:
-    //
-    // <html> may be omitted if first thing inside is not comment
-    // <head> may be omitted if first thing inside is an element
-    // <body> may be omitted if first thing inside is not space, comment, <meta>, <link>, <script>, <style> or <template>
-    // <colgroup> may be omitted if first thing inside is <col>
-    // <tbody> may be omitted if first thing inside is <tr>
-    // An <optgroup> element's end tag may be omitted if the optgroup element is immediately followed by another optgroup element, or if there is no more content in the parent element.
-    // A <colgroup> element's start tag may be omitted if the first thing inside the colgroup element is a col element, and if the element is not immediately preceded by another colgroup element whose end tag has been omitted. (It can't be omitted if the element is empty.)
-    // A <colgroup> element's end tag may be omitted if the colgroup element is not immediately followed by ASCII whitespace or a comment.
-    // A <caption> element's end tag may be omitted if the caption element is not immediately followed by ASCII whitespace or a comment.
-    // A <thead> element's end tag may be omitted if the thead element is immediately followed by a tbody or tfoot element.
-    // A <tbody> element's start tag may be omitted if the first thing inside the tbody element is a tr element, and if the element is not immediately preceded by a tbody, thead, or tfoot element whose end tag has been omitted. (It can't be omitted if the element is empty.)
-    // A <tbody> element's end tag may be omitted if the tbody element is immediately followed by a tbody or tfoot element, or if there is no more content in the parent element.
-    // A <tfoot> element's end tag may be omitted if there is no more content in the parent element.
-    //
-    // <-- However, a start tag must never be omitted if it has any attributes.
+        return $this;
+    }
 
-    return \in_array($tag_name, self::$optional_end_tags, true)
-           ||
-           (
-               $tag_name == 'li'
-               &&
+    /**
+     * @param bool $doRemoveComments
+     *
+     * @return $this
+     */
+    public function doRemoveComments(bool $doRemoveComments = true): self
+    {
+        $this->doRemoveComments = $doRemoveComments;
+
+        return $this;
+    }
+
+    /**
+     * @param bool $doRemoveDefaultAttributes
+     *
+     * @return $this
+     */
+    public function doRemoveDefaultAttributes(bool $doRemoveDefaultAttributes = true): self
+    {
+        $this->doRemoveDefaultAttributes = $doRemoveDefaultAttributes;
+
+        return $this;
+    }
+
+    /**
+     * @param bool $doRemoveDeprecatedAnchorName
+     *
+     * @return $this
+     */
+    public function doRemoveDeprecatedAnchorName(bool $doRemoveDeprecatedAnchorName = true): self
+    {
+        $this->doRemoveDeprecatedAnchorName = $doRemoveDeprecatedAnchorName;
+
+        return $this;
+    }
+
+    /**
+     * @param bool $doRemoveDeprecatedScriptCharsetAttribute
+     *
+     * @return $this
+     */
+    public function doRemoveDeprecatedScriptCharsetAttribute(bool $doRemoveDeprecatedScriptCharsetAttribute = true): self
+    {
+        $this->doRemoveDeprecatedScriptCharsetAttribute = $doRemoveDeprecatedScriptCharsetAttribute;
+
+        return $this;
+    }
+
+    /**
+     * @param bool $doRemoveDeprecatedTypeFromScriptTag
+     *
+     * @return $this
+     */
+    public function doRemoveDeprecatedTypeFromScriptTag(bool $doRemoveDeprecatedTypeFromScriptTag = true): self
+    {
+        $this->doRemoveDeprecatedTypeFromScriptTag = $doRemoveDeprecatedTypeFromScriptTag;
+
+        return $this;
+    }
+
+    /**
+     * @param bool $doRemoveDeprecatedTypeFromStylesheetLink
+     *
+     * @return $this
+     */
+    public function doRemoveDeprecatedTypeFromStylesheetLink(bool $doRemoveDeprecatedTypeFromStylesheetLink = true): self
+    {
+        $this->doRemoveDeprecatedTypeFromStylesheetLink = $doRemoveDeprecatedTypeFromStylesheetLink;
+
+        return $this;
+    }
+
+    /**
+     * @param bool $doRemoveEmptyAttributes
+     *
+     * @return $this
+     */
+    public function doRemoveEmptyAttributes(bool $doRemoveEmptyAttributes = true): self
+    {
+        $this->doRemoveEmptyAttributes = $doRemoveEmptyAttributes;
+
+        return $this;
+    }
+
+    /**
+     * @param bool $doRemoveHttpPrefixFromAttributes
+     *
+     * @return $this
+     */
+    public function doRemoveHttpPrefixFromAttributes(bool $doRemoveHttpPrefixFromAttributes = true): self
+    {
+        $this->doRemoveHttpPrefixFromAttributes = $doRemoveHttpPrefixFromAttributes;
+
+        return $this;
+    }
+
+    /**
+     * @param bool $doRemoveOmittedHtmlTags
+     *
+     * @return $this
+     */
+    public function doRemoveOmittedHtmlTags(bool $doRemoveOmittedHtmlTags = true): self
+    {
+        $this->doRemoveOmittedHtmlTags = $doRemoveOmittedHtmlTags;
+
+        return $this;
+    }
+
+    /**
+     * @param bool $doRemoveOmittedQuotes
+     *
+     * @return $this
+     */
+    public function doRemoveOmittedQuotes(bool $doRemoveOmittedQuotes = true): self
+    {
+        $this->doRemoveOmittedQuotes = $doRemoveOmittedQuotes;
+
+        return $this;
+    }
+
+    /**
+     * @param bool $doRemoveSpacesBetweenTags
+     *
+     * @return $this
+     */
+    public function doRemoveSpacesBetweenTags(bool $doRemoveSpacesBetweenTags = true): self
+    {
+        $this->doRemoveSpacesBetweenTags = $doRemoveSpacesBetweenTags;
+
+        return $this;
+    }
+
+    /**
+     * @param bool $doRemoveValueFromEmptyInput
+     *
+     * @return $this
+     */
+    public function doRemoveValueFromEmptyInput(bool $doRemoveValueFromEmptyInput = true): self
+    {
+        $this->doRemoveValueFromEmptyInput = $doRemoveValueFromEmptyInput;
+
+        return $this;
+    }
+
+    /**
+     * @param bool $doRemoveWhitespaceAroundTags
+     *
+     * @return $this
+     */
+    public function doRemoveWhitespaceAroundTags(bool $doRemoveWhitespaceAroundTags = true): self
+    {
+        $this->doRemoveWhitespaceAroundTags = $doRemoveWhitespaceAroundTags;
+
+        return $this;
+    }
+
+    /**
+     * @param bool $doSortCssClassNames
+     *
+     * @return $this
+     */
+    public function doSortCssClassNames(bool $doSortCssClassNames = true): self
+    {
+        $this->doSortCssClassNames = $doSortCssClassNames;
+
+        return $this;
+    }
+
+    /**
+     * @param bool $doSortHtmlAttributes
+     *
+     * @return $this
+     */
+    public function doSortHtmlAttributes(bool $doSortHtmlAttributes = true): self
+    {
+        $this->doSortHtmlAttributes = $doSortHtmlAttributes;
+
+        return $this;
+    }
+
+    /**
+     * @param bool $doSumUpWhitespace
+     *
+     * @return $this
+     */
+    public function doSumUpWhitespace(bool $doSumUpWhitespace = true): self
+    {
+        $this->doSumUpWhitespace = $doSumUpWhitespace;
+
+        return $this;
+    }
+
+    private function domNodeAttributesToString(\DOMNode $node): string
+    {
+        // Remove quotes around attribute values, when allowed (<p class="foo"> → <p class=foo>)
+        $attr_str = '';
+        if ($node->attributes !== null) {
+            foreach ($node->attributes as $attribute) {
+                $attr_str .= $attribute->name;
+
+                if (
+                    $this->doOptimizeAttributes
+                    &&
+                    isset(self::$booleanAttributes[$attribute->name])
+                ) {
+                    $attr_str .= ' ';
+
+                    continue;
+                }
+
+                $attr_str .= '=';
+
+                // http://www.whatwg.org/specs/web-apps/current-work/multipage/syntax.html#attributes-0
+                $omit_quotes = $this->doRemoveOmittedQuotes
+                               &&
+                               $attribute->value !== ''
+                               &&
+                               \strpos($attribute->name, '____SIMPLE_HTML_DOM__VOKU') !== 0
+                               &&
+                               \strpos($attribute->name, ' ') === false
+                               &&
+                               \preg_match('/["\'=<>` \t\r\n\f]/', $attribute->value) === 0;
+
+                $quoteTmp = '"';
+                if (
+                    !$omit_quotes
+                    &&
+                    \strpos($attribute->value, '"') !== false
+                ) {
+                    $quoteTmp = "'";
+                }
+
+                if (
+                    $this->doOptimizeAttributes
+                    &&
+                    (
+                        $attribute->name === 'srcset'
+                        ||
+                        $attribute->name === 'sizes'
+                    )
+                ) {
+                    $attr_val = \preg_replace(self::$regExSpace, ' ', $attribute->value);
+                } else {
+                    $attr_val = $attribute->value;
+                }
+
+                $attr_str .= ($omit_quotes ? '' : $quoteTmp) . $attr_val . ($omit_quotes ? '' : $quoteTmp);
+                $attr_str .= ' ';
+            }
+        }
+
+        return \trim($attr_str);
+    }
+
+    /**
+     * @param \DOMNode $node
+     *
+     * @return bool
+     */
+    private function domNodeClosingTagOptional(\DOMNode $node): bool
+    {
+        $tag_name = $node->nodeName;
+
+        /** @var \DOMNode|null $parent_node - false-positive error from phpstan */
+        $parent_node = $node->parentNode;
+
+        if ($parent_node) {
+            $parent_tag_name = $parent_node->nodeName;
+        } else {
+            $parent_tag_name = null;
+        }
+
+        $nextSibling = $this->getNextSiblingOfTypeDOMElement($node);
+
+        // https://html.spec.whatwg.org/multipage/syntax.html#syntax-tag-omission
+
+        // Implemented:
+        //
+        // A <p> element's end tag may be omitted if the p element is immediately followed by an address, article, aside, blockquote, details, div, dl, fieldset, figcaption, figure, footer, form, h1, h2, h3, h4, h5, h6, header, hgroup, hr, main, menu, nav, ol, p, pre, section, table, or ul element, or if there is no more content in the parent element and the parent element is an HTML element that is not an a, audio, del, ins, map, noscript, or video element, or an autonomous custom element.
+        // An <li> element's end tag may be omitted if the li element is immediately followed by another li element or if there is no more content in the parent element.
+        // A <td> element's end tag may be omitted if the td element is immediately followed by a td or th element, or if there is no more content in the parent element.
+        // An <option> element's end tag may be omitted if the option element is immediately followed by another option element, or if it is immediately followed by an optgroup element, or if there is no more content in the parent element.
+        // A <tr> element's end tag may be omitted if the tr element is immediately followed by another tr element, or if there is no more content in the parent element.
+        // A <th> element's end tag may be omitted if the th element is immediately followed by a td or th element, or if there is no more content in the parent element.
+        // A <dt> element's end tag may be omitted if the dt element is immediately followed by another dt element or a dd element.
+        // A <dd> element's end tag may be omitted if the dd element is immediately followed by another dd element or a dt element, or if there is no more content in the parent element.
+        // An <rp> element's end tag may be omitted if the rp element is immediately followed by an rt or rp element, or if there is no more content in the parent element.
+
+        /**
+         * @noinspection TodoComment
+         *
+         * TODO: Not Implemented
+         */
+        //
+        // <html> may be omitted if first thing inside is not comment
+        // <head> may be omitted if first thing inside is an element
+        // <body> may be omitted if first thing inside is not space, comment, <meta>, <link>, <script>, <style> or <template>
+        // <colgroup> may be omitted if first thing inside is <col>
+        // <tbody> may be omitted if first thing inside is <tr>
+        // An <optgroup> element's end tag may be omitted if the optgroup element is immediately followed by another optgroup element, or if there is no more content in the parent element.
+        // A <colgroup> element's start tag may be omitted if the first thing inside the colgroup element is a col element, and if the element is not immediately preceded by another colgroup element whose end tag has been omitted. (It can't be omitted if the element is empty.)
+        // A <colgroup> element's end tag may be omitted if the colgroup element is not immediately followed by ASCII whitespace or a comment.
+        // A <caption> element's end tag may be omitted if the caption element is not immediately followed by ASCII whitespace or a comment.
+        // A <thead> element's end tag may be omitted if the thead element is immediately followed by a tbody or tfoot element.
+        // A <tbody> element's start tag may be omitted if the first thing inside the tbody element is a tr element, and if the element is not immediately preceded by a tbody, thead, or tfoot element whose end tag has been omitted. (It can't be omitted if the element is empty.)
+        // A <tbody> element's end tag may be omitted if the tbody element is immediately followed by a tbody or tfoot element, or if there is no more content in the parent element.
+        // A <tfoot> element's end tag may be omitted if there is no more content in the parent element.
+        //
+        // <-- However, a start tag must never be omitted if it has any attributes.
+
+        return \in_array($tag_name, self::$optional_end_tags, true)
+               ||
                (
-                   $nextSibling === null
-                   ||
-                   (
-                       $nextSibling instanceof \DOMElement
-                       &&
-                       $nextSibling->tagName == 'li'
-                   )
-               )
-           )
-           ||
-           (
-               (
-                   $tag_name == 'rp'
-               )
-               &&
-               (
-                   $nextSibling === null
-                   ||
-                   (
-                       $nextSibling instanceof \DOMElement
-                       &&
-                       (
-                           $nextSibling->tagName == 'rp'
-                           ||
-                           $nextSibling->tagName == 'rt'
-                       )
-                   )
-               )
-           )
-           ||
-           (
-               $tag_name == 'tr'
-               &&
-               (
-                   $nextSibling === null
-                   ||
-                   (
-                       $nextSibling instanceof \DOMElement
-                       &&
-                       $nextSibling->tagName == 'tr'
-                   )
-               )
-           )
-           ||
-           (
-               (
-                   $tag_name == 'td'
-                   ||
-                   $tag_name == 'th'
-               )
-               &&
-               (
-                   $nextSibling === null
-                   ||
-                   (
-                       $nextSibling instanceof \DOMElement
-                       &&
-                       (
-                           $nextSibling->tagName == 'td'
-                           ||
-                           $nextSibling->tagName == 'th'
-                       )
-                   )
-               )
-           )
-           ||
-           (
-               (
-                   $tag_name == 'dd'
-                   ||
-                   $tag_name == 'dt'
-               )
-               &&
-               (
+                   $tag_name === 'li'
+                   &&
                    (
                        $nextSibling === null
-                       &&
-                       $tag_name == 'dd'
-                   )
-                   ||
-                   (
-                       $nextSibling instanceof \DOMElement
-                       &&
+                       ||
                        (
-                           $nextSibling->tagName == 'dd'
-                           ||
-                           $nextSibling->tagName == 'dt'
-                       )
-                   )
-               )
-           )
-           ||
-           (
-               $tag_name == 'option'
-               &&
-               (
-                   $nextSibling === null
-                   ||
-                   (
-                       $nextSibling instanceof \DOMElement
-                       &&
-                       (
-                           $nextSibling->tagName == 'option'
-                           ||
-                           $nextSibling->tagName == 'optgroup'
-                       )
-                   )
-               )
-           )
-           ||
-           (
-               $tag_name == 'p'
-               &&
-               (
-                   (
-                       $nextSibling === null
-                       &&
-                       (
-                           $node->parentNode !== null
+                           $nextSibling instanceof \DOMElement
                            &&
-                           !\in_array(
-                               $node->parentNode->nodeName,
+                           $nextSibling->tagName === 'li'
+                       )
+                   )
+               )
+               ||
+               (
+                   $tag_name === 'rp'
+                   &&
+                   (
+                       $nextSibling === null
+                       ||
+                       (
+                           $nextSibling instanceof \DOMElement
+                           &&
+                           (
+                               $nextSibling->tagName === 'rp'
+                               ||
+                               $nextSibling->tagName === 'rt'
+                           )
+                       )
+                   )
+               )
+               ||
+               (
+                   $tag_name === 'tr'
+                   &&
+                   (
+                       $nextSibling === null
+                       ||
+                       (
+                           $nextSibling instanceof \DOMElement
+                           &&
+                           $nextSibling->tagName === 'tr'
+                       )
+                   )
+               )
+               ||
+               (
+                   $tag_name === 'source'
+                   &&
+                   (
+                       $parent_tag_name === 'audio'
+                       ||
+                       $parent_tag_name === 'video'
+                       ||
+                       $parent_tag_name === 'picture'
+                       ||
+                       $parent_tag_name === 'source'
+                   )
+                   &&
+                   (
+                       $nextSibling === null
+                       ||
+                       (
+                           $nextSibling instanceof \DOMElement
+                           &&
+                           $nextSibling->tagName === 'source'
+                       )
+                   )
+               )
+               ||
+               (
+                   (
+                       $tag_name === 'td'
+                       ||
+                       $tag_name === 'th'
+                   )
+                   &&
+                   (
+                       $nextSibling === null
+                       ||
+                       (
+                           $nextSibling instanceof \DOMElement
+                           &&
+                           (
+                               $nextSibling->tagName === 'td'
+                               ||
+                               $nextSibling->tagName === 'th'
+                           )
+                       )
+                   )
+               )
+               ||
+               (
+                   (
+                       $tag_name === 'dd'
+                       ||
+                       $tag_name === 'dt'
+                   )
+                   &&
+                   (
+                       (
+                           $nextSibling === null
+                           &&
+                           $tag_name === 'dd'
+                       )
+                       ||
+                       (
+                           $nextSibling instanceof \DOMElement
+                           &&
+                           (
+                               $nextSibling->tagName === 'dd'
+                               ||
+                               $nextSibling->tagName === 'dt'
+                           )
+                       )
+                   )
+               )
+               ||
+               (
+                   $tag_name === 'option'
+                   &&
+                   (
+                       $nextSibling === null
+                       ||
+                       (
+                           $nextSibling instanceof \DOMElement
+                           &&
+                           (
+                               $nextSibling->tagName === 'option'
+                               ||
+                               $nextSibling->tagName === 'optgroup'
+                           )
+                       )
+                   )
+               )
+               ||
+               (
+                   $tag_name === 'p'
+                   &&
+                   (
+                       (
+                           $nextSibling === null
+                           &&
+                           (
+                               $node->parentNode !== null
+                               &&
+                               !\in_array(
+                                   $node->parentNode->nodeName,
+                                   [
+                                       'a',
+                                       'audio',
+                                       'del',
+                                       'ins',
+                                       'map',
+                                       'noscript',
+                                       'video',
+                                   ],
+                                   true
+                               )
+                           )
+                       )
+                       ||
+                       (
+                           $nextSibling instanceof \DOMElement
+                           &&
+                           \in_array(
+                               $nextSibling->tagName,
                                [
-                                   'a',
-                                   'audio',
-                                   'del',
-                                   'ins',
-                                   'map',
-                                   'noscript',
-                                   'video',
+                                   'address',
+                                   'article',
+                                   'aside',
+                                   'blockquote',
+                                   'dir',
+                                   'div',
+                                   'dl',
+                                   'fieldset',
+                                   'footer',
+                                   'form',
+                                   'h1',
+                                   'h2',
+                                   'h3',
+                                   'h4',
+                                   'h5',
+                                   'h6',
+                                   'header',
+                                   'hgroup',
+                                   'hr',
+                                   'menu',
+                                   'nav',
+                                   'ol',
+                                   'p',
+                                   'pre',
+                                   'section',
+                                   'table',
+                                   'ul',
                                ],
                                true
                            )
                        )
                    )
-                   ||
-                   (
-                       $nextSibling instanceof \DOMElement
-                       &&
-                       \in_array(
-                           $nextSibling->tagName,
-                           [
-                               'address',
-                               'article',
-                               'aside',
-                               'blockquote',
-                               'dir',
-                               'div',
-                               'dl',
-                               'fieldset',
-                               'footer',
-                               'form',
-                               'h1',
-                               'h2',
-                               'h3',
-                               'h4',
-                               'h5',
-                               'h6',
-                               'header',
-                               'hgroup',
-                               'hr',
-                               'menu',
-                               'nav',
-                               'ol',
-                               'p',
-                               'pre',
-                               'section',
-                               'table',
-                               'ul',
-                           ],
-                           true
-                       )
-                   )
-               )
-           );
-  }
+               );
+    }
 
-  protected function domNodeToString(\DOMNode $node): string
-  {
-    // init
-    $html = '';
-    $emptyStringTmp = '';
-
-    foreach ($node->childNodes as $child) {
-
-      if ($emptyStringTmp === 'is_empty') {
-        $emptyStringTmp = 'last_was_empty';
-      } else {
+    protected function domNodeToString(\DOMNode $node): string
+    {
+        // init
+        $html = '';
         $emptyStringTmp = '';
-      }
 
-      if ($child instanceof \DOMDocumentType) {
-
-        // add the doc-type only if it wasn't generated by DomDocument
-        if ($this->withDocType !== true) {
-          continue;
-        }
-
-        if ($child->name) {
-
-          if (!$child->publicId && $child->systemId) {
-            $tmpTypeSystem = 'SYSTEM';
-            $tmpTypePublic = '';
-          } else {
-            $tmpTypeSystem = '';
-            $tmpTypePublic = 'PUBLIC';
-          }
-
-          $html .= '<!DOCTYPE ' . $child->name . ''
-                   . ($child->publicId ? ' ' . $tmpTypePublic . ' "' . $child->publicId . '"' : '')
-                   . ($child->systemId ? ' ' . $tmpTypeSystem . ' "' . $child->systemId . '"' : '')
-                   . '>';
-        }
-
-      } elseif ($child instanceof \DOMElement) {
-
-        $html .= \rtrim('<' . $child->tagName . ' ' . $this->domNodeAttributesToString($child));
-        $html .= '>' . $this->domNodeToString($child);
-
-        if (
-            $this->doRemoveOmittedHtmlTags === false
-            ||
-            !$this->domNodeClosingTagOptional($child)
-        ) {
-          $html .= '</' . $child->tagName . '>';
-        }
-
-        if ($this->doRemoveWhitespaceAroundTags === false) {
-          if (
-              $child->nextSibling instanceof \DOMText
-              &&
-              $child->nextSibling->wholeText === ' '
-          ) {
-            if (
-                $emptyStringTmp !== 'last_was_empty'
-                &&
-                substr($html, -1) !== ' '
-            ) {
-              $html .= ' ';
+        foreach ($node->childNodes as $child) {
+            if ($emptyStringTmp === 'is_empty') {
+                $emptyStringTmp = 'last_was_empty';
+            } else {
+                $emptyStringTmp = '';
             }
-            $emptyStringTmp = 'is_empty';
-          }
-        }
 
-      } elseif ($child instanceof \DOMText) {
+            if ($child instanceof \DOMDocumentType) {
+                // add the doc-type only if it wasn't generated by DomDocument
+                if (!$this->withDocType) {
+                    continue;
+                }
 
-        if ($child->isElementContentWhitespace()) {
-          if (
-              $child->previousSibling !== null
-              &&
-              $child->nextSibling !== null
-          ) {
-            if (
-                $emptyStringTmp !== 'last_was_empty'
-                &&
-                substr($html, -1) !== ' '
-            ) {
-              $html .= ' ';
+                if ($child->name) {
+                    if (!$child->publicId && $child->systemId) {
+                        $tmpTypeSystem = 'SYSTEM';
+                        $tmpTypePublic = '';
+                    } else {
+                        $tmpTypeSystem = '';
+                        $tmpTypePublic = 'PUBLIC';
+                    }
+
+                    $html .= '<!DOCTYPE ' . $child->name . ''
+                             . ($child->publicId ? ' ' . $tmpTypePublic . ' "' . $child->publicId . '"' : '')
+                             . ($child->systemId ? ' ' . $tmpTypeSystem . ' "' . $child->systemId . '"' : '')
+                             . '>';
+                }
+            } elseif ($child instanceof \DOMElement) {
+                $html .= \rtrim('<' . $child->tagName . ' ' . $this->domNodeAttributesToString($child));
+                $html .= '>' . $this->domNodeToString($child);
+
+                if (
+                    !$this->doRemoveOmittedHtmlTags
+                    ||
+                    !$this->domNodeClosingTagOptional($child)
+                ) {
+                    $html .= '</' . $child->tagName . '>';
+                }
+
+                if (!$this->doRemoveWhitespaceAroundTags) {
+                    /** @noinspection NestedPositiveIfStatementsInspection */
+                    if (
+                        $child->nextSibling instanceof \DOMText
+                        &&
+                        $child->nextSibling->wholeText === ' '
+                    ) {
+                        if (
+                            $emptyStringTmp !== 'last_was_empty'
+                            &&
+                            \substr($html, -1) !== ' '
+                        ) {
+                            $html = \rtrim($html);
+
+                            if (
+                                $child->parentNode
+                                &&
+                                $child->parentNode->nodeName !== 'head'
+                            ) {
+                                $html .= ' ';
+                            }
+                        }
+                        $emptyStringTmp = 'is_empty';
+                    }
+                }
+            } elseif ($child instanceof \DOMText) {
+                if ($child->isElementContentWhitespace()) {
+                    if (
+                        $child->previousSibling !== null
+                        &&
+                        $child->nextSibling !== null
+                    ) {
+                        if (
+                            (
+                                $child->wholeText
+                                &&
+                                strpos($child->wholeText, ' ') !== false
+                            )
+                            ||
+                            (
+                                $emptyStringTmp !== 'last_was_empty'
+                                &&
+                                \substr($html, -1) !== ' '
+                            )
+                        ) {
+                            $html = \rtrim($html);
+
+                            if (
+                                $child->parentNode
+                                &&
+                                $child->parentNode->nodeName !== 'head'
+                            ) {
+                                $html .= ' ';
+                            }
+                        }
+                        $emptyStringTmp = 'is_empty';
+                    }
+                } else {
+                    $html .= $child->wholeText;
+                }
+            } elseif ($child instanceof \DOMComment) {
+                $html .= '<!--' . $child->textContent . '-->';
             }
-            $emptyStringTmp = 'is_empty';
-          }
-
-        } else {
-
-          $html .= $child->wholeText;
-
         }
 
-      } elseif ($child instanceof \DOMComment) {
-
-        $html .= $child->wholeText;
-
-      }
+        return $html;
     }
 
-    return $html;
-  }
-
-  /**
-   * @param \DOMNode $node
-   *
-   * @return \DOMNode|null
-   */
-  protected function getNextSiblingOfTypeDOMElement(\DOMNode $node)
-  {
-    do {
-      $node = $node->nextSibling;
-    } while (!($node === null || $node instanceof \DOMElement));
-
-    return $node;
-  }
-
-  /**
-   * Check if the current string is an conditional comment.
-   *
-   * INFO: since IE >= 10 conditional comment are not working anymore
-   *
-   * <!--[if expression]> HTML <![endif]-->
-   * <![if expression]> HTML <![endif]>
-   *
-   * @param string $comment
-   *
-   * @return bool
-   */
-  private function isConditionalComment($comment): bool
-  {
-    if (preg_match('/^\[if [^\]]+\]/', $comment)) {
-      return true;
+    /**
+     * @return array
+     */
+    public function getDomainsToRemoveHttpPrefixFromAttributes(): array
+    {
+        return $this->domainsToRemoveHttpPrefixFromAttributes;
     }
 
-    if (preg_match('/\[endif\]$/', $comment)) {
-      return true;
+    /**
+     * @return bool
+     */
+    public function isDoOptimizeAttributes(): bool
+    {
+        return $this->doOptimizeAttributes;
     }
 
-    return false;
-  }
-
-  /**
-   * @param string $html
-   * @param bool   $decodeUtf8Specials <p>Use this only in special cases, e.g. for PHP 5.3</p>
-   *
-   * @return string
-   */
-  public function minify($html, $decodeUtf8Specials = false): string
-  {
-    $html = (string)$html;
-    if (!isset($html[0])) {
-      return '';
+    /**
+     * @return bool
+     */
+    public function isDoOptimizeViaHtmlDomParser(): bool
+    {
+        return $this->doOptimizeViaHtmlDomParser;
     }
 
-    $html = \trim($html);
-    if (!$html) {
-      return '';
+    /**
+     * @return bool
+     */
+    public function isDoRemoveComments(): bool
+    {
+        return $this->doRemoveComments;
     }
 
-    // init
-    static $CACHE_SELF_CLOSING_TAGS = null;
-    if ($CACHE_SELF_CLOSING_TAGS === null) {
-      $CACHE_SELF_CLOSING_TAGS = \implode('|', self::$selfClosingTags);
+    /**
+     * @return bool
+     */
+    public function isDoRemoveDefaultAttributes(): bool
+    {
+        return $this->doRemoveDefaultAttributes;
     }
 
-    // reset
-    $this->protectedChildNodes = [];
-
-    // save old content
-    $origHtml = $html;
-    $origHtmlLength = \strlen($html);
-
-    // -------------------------------------------------------------------------
-    // Minify the HTML via "HtmlDomParser"
-    // -------------------------------------------------------------------------
-
-    if ($this->doOptimizeViaHtmlDomParser === true) {
-      $html = $this->minifyHtmlDom($html, $decodeUtf8Specials);
+    /**
+     * @return bool
+     */
+    public function isDoRemoveDeprecatedAnchorName(): bool
+    {
+        return $this->doRemoveDeprecatedAnchorName;
     }
 
-    // -------------------------------------------------------------------------
-    // Trim whitespace from html-string. [protected html is still protected]
-    // -------------------------------------------------------------------------
-
-    // Remove extra white-space(s) between HTML attribute(s)
-    $html = (string)\preg_replace_callback(
-        '#<([^\/\s<>!]+)(?:\s+([^<>]*?)\s*|\s*)(\/?)>#',
-        function ($matches) {
-          return '<' . $matches[1] . (string)\preg_replace('#([^\s=]+)(\=([\'"]?)(.*?)\3)?(\s+|$)#s', ' $1$2', $matches[2]) . $matches[3] . '>';
-        },
-        $html
-    );
-
-    if ($this->doRemoveSpacesBetweenTags === true) {
-      // Remove spaces that are between > and <
-      $html = (string)\preg_replace('/(>) (<)/', '>$2', $html);
+    /**
+     * @return bool
+     */
+    public function isDoRemoveDeprecatedScriptCharsetAttribute(): bool
+    {
+        return $this->doRemoveDeprecatedScriptCharsetAttribute;
     }
 
-    // -------------------------------------------------------------------------
-    // Restore protected HTML-code.
-    // -------------------------------------------------------------------------
-
-    $html = (string)\preg_replace_callback(
-        '/<(?<element>' . $this->protectedChildNodesHelper . ')(?<attributes> [^>]*)?>(?<value>.*?)<\/' . $this->protectedChildNodesHelper . '>/',
-        [$this, 'restoreProtectedHtml'],
-        $html
-    );
-
-    // -------------------------------------------------------------------------
-    // Restore protected HTML-entities.
-    // -------------------------------------------------------------------------
-
-    if ($this->doOptimizeViaHtmlDomParser === true) {
-      $html = HtmlDomParser::putReplacedBackToPreserveHtmlEntities($html);
+    /**
+     * @return bool
+     */
+    public function isDoRemoveDeprecatedTypeFromScriptTag(): bool
+    {
+        return $this->doRemoveDeprecatedTypeFromScriptTag;
     }
 
-    // ------------------------------------
-    // Final clean-up
-    // ------------------------------------
-
-    $html = \str_replace(
-        [
-            'html>' . "\n",
-            "\n" . '<html',
-            'html/>' . "\n",
-            "\n" . '</html',
-            'head>' . "\n",
-            "\n" . '<head',
-            'head/>' . "\n",
-            "\n" . '</head',
-        ],
-        [
-            'html>',
-            '<html',
-            'html/>',
-            '</html',
-            'head>',
-            '<head',
-            'head/>',
-            '</head',
-        ],
-        $html
-    );
-
-    // self closing tags, don't need a trailing slash ...
-    $replace = [];
-    $replacement = [];
-    foreach (self::$selfClosingTags as $selfClosingTag) {
-      $replace[] = '<' . $selfClosingTag . '/>';
-      $replacement[] = '<' . $selfClosingTag . '>';
-      $replace[] = '<' . $selfClosingTag . ' />';
-      $replacement[] = '<' . $selfClosingTag . '>';
-    }
-    $html = \str_replace(
-        $replace,
-        $replacement,
-        $html
-    );
-
-    $html = (string)\preg_replace('#<\b(' . $CACHE_SELF_CLOSING_TAGS . ')([^>]+)><\/\b\1>#', '<\\1\\2>', $html);
-
-    // ------------------------------------
-    // check if compression worked
-    // ------------------------------------
-
-    if ($origHtmlLength < \strlen($html)) {
-      $html = $origHtml;
+    /**
+     * @return bool
+     */
+    public function isDoRemoveDeprecatedTypeFromStylesheetLink(): bool
+    {
+        return $this->doRemoveDeprecatedTypeFromStylesheetLink;
     }
 
-    return $html;
-  }
-
-  /**
-   * @param $html
-   * @param $decodeUtf8Specials
-   *
-   * @return string
-   */
-  private function minifyHtmlDom($html, $decodeUtf8Specials): string
-  {
-    // init dom
-    $dom = new HtmlDomParser();
-    $dom->getDocument()->preserveWhiteSpace = false; // remove redundant white space
-    $dom->getDocument()->formatOutput = false; // do not formats output with indentation
-
-    // load dom
-    $dom->loadHtml($html);
-
-    $this->withDocType = (\stripos(\ltrim($html), '<!DOCTYPE') === 0);
-
-    // -------------------------------------------------------------------------
-    // Protect HTML tags and conditional comments.
-    // -------------------------------------------------------------------------
-
-    $dom = $this->protectTags($dom);
-
-    // -------------------------------------------------------------------------
-    // Remove default HTML comments. [protected html is still protected]
-    // -------------------------------------------------------------------------
-
-    if ($this->doRemoveComments === true) {
-      $dom = $this->removeComments($dom);
+    /**
+     * @return bool
+     */
+    public function isDoRemoveEmptyAttributes(): bool
+    {
+        return $this->doRemoveEmptyAttributes;
     }
 
-    // -------------------------------------------------------------------------
-    // Sum-Up extra whitespace from the Dom. [protected html is still protected]
-    // -------------------------------------------------------------------------
-
-    if ($this->doSumUpWhitespace === true) {
-      $dom = $this->sumUpWhitespace($dom);
+    /**
+     * @return bool
+     */
+    public function isDoRemoveHttpPrefixFromAttributes(): bool
+    {
+        return $this->doRemoveHttpPrefixFromAttributes;
     }
 
-    foreach ($dom->find('*') as $element) {
-
-      // -------------------------------------------------------------------------
-      // Optimize html attributes. [protected html is still protected]
-      // -------------------------------------------------------------------------
-
-      if ($this->doOptimizeAttributes === true) {
-        $this->optimizeAttributes($element);
-      }
-
-      // -------------------------------------------------------------------------
-      // Remove whitespace around tags. [protected html is still protected]
-      // -------------------------------------------------------------------------
-
-      if ($this->doRemoveWhitespaceAroundTags === true) {
-        $this->removeWhitespaceAroundTags($element);
-      }
+    /**
+     * @return bool
+     */
+    public function isDoRemoveOmittedHtmlTags(): bool
+    {
+        return $this->doRemoveOmittedHtmlTags;
     }
 
-    // -------------------------------------------------------------------------
-    // Convert the Dom into a string.
-    // -------------------------------------------------------------------------
-
-    $html = $dom->fixHtmlOutput(
-        $this->domNodeToString($dom->getDocument()),
-        $decodeUtf8Specials
-    );
-
-    return $html;
-  }
-
-  /**
-   * Sort HTML-Attributes, so that gzip can do better work and remove some default attributes...
-   *
-   * @param SimpleHtmlDom $element
-   *
-   * @return bool
-   */
-  private function optimizeAttributes(SimpleHtmlDom $element): bool
-  {
-    $attributes = $element->getAllAttributes();
-    if ($attributes === null) {
-      return false;
+    /**
+     * @return bool
+     */
+    public function isDoRemoveOmittedQuotes(): bool
+    {
+        return $this->doRemoveOmittedQuotes;
     }
 
-    $attrs = [];
-    foreach ((array)$attributes as $attrName => $attrValue) {
-
-      // -------------------------------------------------------------------------
-      // Remove optional "http:"-prefix from attributes.
-      // -------------------------------------------------------------------------
-
-      if ($this->doRemoveHttpPrefixFromAttributes === true) {
-        if (
-            ($attrName === 'href' || $attrName === 'src' || $attrName === 'action')
-            &&
-            !(isset($attributes['rel']) && $attributes['rel'] === 'external')
-            &&
-            !(isset($attributes['target']) && $attributes['target'] === '_blank')
-        ) {
-          $attrValue = \str_replace('http://', '//', $attrValue);
-        }
-      }
-
-      if ($this->removeAttributeHelper($element->tag, $attrName, $attrValue, $attributes)) {
-        $element->{$attrName} = null;
-        continue;
-      }
-
-      // -------------------------------------------------------------------------
-      // Sort css-class-names, for better gzip results.
-      // -------------------------------------------------------------------------
-
-      if ($this->doSortCssClassNames === true) {
-        $attrValue = $this->sortCssClassNames($attrName, $attrValue);
-      }
-
-      if ($this->doSortHtmlAttributes === true) {
-        $attrs[$attrName] = $attrValue;
-        $element->{$attrName} = null;
-      }
+    /**
+     * @return bool
+     */
+    public function isDoRemoveSpacesBetweenTags(): bool
+    {
+        return $this->doRemoveSpacesBetweenTags;
     }
 
-    // -------------------------------------------------------------------------
-    // Sort html-attributes, for better gzip results.
-    // -------------------------------------------------------------------------
-
-    if ($this->doSortHtmlAttributes === true) {
-      \ksort($attrs);
-      foreach ($attrs as $attrName => $attrValue) {
-        $attrValue = HtmlDomParser::replaceToPreserveHtmlEntities($attrValue);
-        $element->setAttribute($attrName, $attrValue, true);
-      }
+    /**
+     * @return bool
+     */
+    public function isDoRemoveValueFromEmptyInput(): bool
+    {
+        return $this->doRemoveValueFromEmptyInput;
     }
 
-    return true;
-  }
-
-  /**
-   * Prevent changes of inline "styles" and "scripts".
-   *
-   * @param HtmlDomParser $dom
-   *
-   * @return HtmlDomParser
-   */
-  private function protectTags(HtmlDomParser $dom): HtmlDomParser
-  {
-    // init
-    $counter = 0;
-
-    foreach ($dom->find('script, style') as $element) {
-
-      // skip external links
-      if ($element->tag === 'script' || $element->tag === 'style') {
-        $attributes = $element->getAllAttributes();
-        if (isset($attributes['src'])) {
-          continue;
-        }
-      }
-
-      $this->protectedChildNodes[$counter] = $element->text();
-      $element->getNode()->nodeValue = '<' . $this->protectedChildNodesHelper . ' data-' . $this->protectedChildNodesHelper . '="' . $counter . '"></' . $this->protectedChildNodesHelper . '>';
-
-      ++$counter;
+    /**
+     * @return bool
+     */
+    public function isDoRemoveWhitespaceAroundTags(): bool
+    {
+        return $this->doRemoveWhitespaceAroundTags;
     }
 
-    $dom->getDocument()->normalizeDocument();
-
-    foreach ($dom->find('//comment()') as $element) {
-      $text = $element->text();
-
-      // skip normal comments
-      if ($this->isConditionalComment($text) === false) {
-        continue;
-      }
-
-      $this->protectedChildNodes[$counter] = '<!--' . $text . '-->';
-
-      /* @var $node \DOMComment */
-      $node = $element->getNode();
-      $child = new \DOMText('<' . $this->protectedChildNodesHelper . ' data-' . $this->protectedChildNodesHelper . '="' . $counter . '"></' . $this->protectedChildNodesHelper . '>');
-      $element->getNode()->parentNode->replaceChild($child, $node);
-
-      ++$counter;
+    /**
+     * @return bool
+     */
+    public function isDoSortCssClassNames(): bool
+    {
+        return $this->doSortCssClassNames;
     }
 
-    $dom->getDocument()->normalizeDocument();
-
-    return $dom;
-  }
-
-  /**
-   * Check if the attribute can be removed.
-   *
-   * @param string $tag
-   * @param string $attrName
-   * @param string $attrValue
-   * @param array  $allAttr
-   *
-   * @return bool
-   */
-  private function removeAttributeHelper($tag, $attrName, $attrValue, $allAttr): bool
-  {
-    // remove defaults
-    if ($this->doRemoveDefaultAttributes === true) {
-
-      if ($tag === 'script' && $attrName === 'language' && $attrValue === 'javascript') {
-        return true;
-      }
-
-      if ($tag === 'form' && $attrName === 'method' && $attrValue === 'get') {
-        return true;
-      }
-
-      if ($tag === 'input' && $attrName === 'type' && $attrValue === 'text') {
-        return true;
-      }
-
-      if ($tag === 'area' && $attrName === 'shape' && $attrValue === 'rect') {
-        return true;
-      }
+    /**
+     * @return bool
+     */
+    public function isDoSortHtmlAttributes(): bool
+    {
+        return $this->doSortHtmlAttributes;
     }
 
-    // remove deprecated charset-attribute (the browser will use the charset from the HTTP-Header, anyway)
-    if ($this->doRemoveDeprecatedScriptCharsetAttribute === true) {
-      if ($tag === 'script' && $attrName === 'charset' && !isset($allAttr['src'])) {
-        return true;
-      }
+    /**
+     * @return bool
+     */
+    public function isDoSumUpWhitespace(): bool
+    {
+        return $this->doSumUpWhitespace;
     }
 
-    // remove deprecated anchor-jump
-    if ($this->doRemoveDeprecatedAnchorName === true) {
-      if ($tag === 'a' && $attrName === 'name' && isset($allAttr['id']) && $allAttr['id'] === $attrValue) {
-        return true;
-      }
-    }
-
-    // remove "type=text/css" for css links
-    if ($this->doRemoveDeprecatedTypeFromStylesheetLink === true) {
-      if ($tag === 'link' && $attrName === 'type' && $attrValue === 'text/css' && isset($allAttr['rel']) && $allAttr['rel'] === 'stylesheet') {
-        return true;
-      }
-    }
-
-    // remove deprecated script-mime-types
-    if ($this->doRemoveDeprecatedTypeFromScriptTag === true) {
-      if ($tag === 'script' && $attrName === 'type' && isset($allAttr['src'], self::$executableScriptsMimeTypes[$attrValue])) {
-        return true;
-      }
-    }
-
-    // remove 'value=""' from <input type="text">
-    if ($this->doRemoveValueFromEmptyInput === true) {
-      if ($tag === 'input' && $attrName === 'value' && $attrValue === '' && isset($allAttr['type']) && $allAttr['type'] === 'text') {
-        return true;
-      }
-    }
-
-    // remove some empty attributes
-    if ($this->doRemoveEmptyAttributes === true) {
-      if (\trim($attrValue) === '' && \preg_match('/^(?:class|id|style|title|lang|dir|on(?:focus|blur|change|click|dblclick|mouse(?:down|up|over|move|out)|key(?:press|down|up)))$/', $attrName)) {
-        return true;
-      }
-    }
-
-    return false;
-  }
-
-  /**
-   * Remove comments in the dom.
-   *
-   * @param HtmlDomParser $dom
-   *
-   * @return HtmlDomParser
-   */
-  private function removeComments(HtmlDomParser $dom): HtmlDomParser
-  {
-    foreach ($dom->find('//comment()') as $commentWrapper) {
-      $comment = $commentWrapper->getNode();
-      $val = $comment->nodeValue;
-      if (\strpos($val, '[') === false) {
-        $comment->parentNode->removeChild($comment);
-      }
-    }
-
-    $dom->getDocument()->normalizeDocument();
-
-    return $dom;
-  }
-
-  /**
-   * Trim tags in the dom.
-   *
-   * @param SimpleHtmlDom $element
-   *
-   * @return void
-   */
-  private function removeWhitespaceAroundTags(SimpleHtmlDom $element)
-  {
-    if (isset(self::$trimWhitespaceFromTags[$element->tag])) {
-      $node = $element->getNode();
-
-      $candidates = [];
-      if ($node->childNodes->length > 0) {
-        $candidates[] = $node->firstChild;
-        $candidates[] = $node->lastChild;
-        $candidates[] = $node->previousSibling;
-        $candidates[] = $node->nextSibling;
-      }
-
-      foreach ($candidates as &$candidate) {
-        if ($candidate === null) {
-          continue;
+    /**
+     * @param string $html
+     * @param bool   $multiDecodeNewHtmlEntity
+     *
+     * @return string
+     */
+    public function minify($html, $multiDecodeNewHtmlEntity = false): string
+    {
+        $html = (string) $html;
+        if (!isset($html[0])) {
+            return '';
         }
 
-        if ($candidate->nodeType === 3) {
-          $candidate->nodeValue = \preg_replace(self::$regExSpace, ' ', $candidate->nodeValue);
+        $html = \trim($html);
+        if (!$html) {
+            return '';
         }
-      }
-    }
-  }
 
-  /**
-   * Callback function for preg_replace_callback use.
-   *
-   * @param array $matches PREG matches
-   *
-   * @return string
-   */
-  private function restoreProtectedHtml($matches): string
-  {
-    \preg_match('/.*"(?<id>\d*)"/', $matches['attributes'], $matchesInner);
+        // reset
+        $this->protectedChildNodes = [];
 
-    $html = '';
-    if (isset($this->protectedChildNodes[$matchesInner['id']])) {
-      $html .= $this->protectedChildNodes[$matchesInner['id']];
-    }
+        // save old content
+        $origHtml = $html;
+        $origHtmlLength = \strlen($html);
 
-    return $html;
-  }
+        // -------------------------------------------------------------------------
+        // Minify the HTML via "HtmlDomParser"
+        // -------------------------------------------------------------------------
 
-  /**
-   * @param array $domainsToRemoveHttpPrefixFromAttributes
-   *
-   * @return $this
-   */
-  public function setDomainsToRemoveHttpPrefixFromAttributes($domainsToRemoveHttpPrefixFromAttributes)
-  {
-    $this->domainsToRemoveHttpPrefixFromAttributes = $domainsToRemoveHttpPrefixFromAttributes;
-
-    return $this;
-  }
-
-  /**
-   * @param $attrName
-   * @param $attrValue
-   *
-   * @return string
-   */
-  private function sortCssClassNames($attrName, $attrValue): string
-  {
-    if ($attrName !== 'class' || !$attrValue) {
-      return $attrValue;
-    }
-
-    $classes = \array_unique(
-        \explode(' ', $attrValue)
-    );
-    \sort($classes);
-
-    $attrValue = '';
-    foreach ($classes as $class) {
-
-      if (!$class) {
-        continue;
-      }
-
-      $attrValue .= \trim($class) . ' ';
-    }
-    $attrValue = \trim($attrValue);
-
-    return $attrValue;
-  }
-
-  /**
-   * Sum-up extra whitespace from dom-nodes.
-   *
-   * @param HtmlDomParser $dom
-   *
-   * @return HtmlDomParser
-   */
-  private function sumUpWhitespace(HtmlDomParser $dom): HtmlDomParser
-  {
-    $textnodes = $dom->find('//text()');
-    foreach ($textnodes as $textnodeWrapper) {
-      /* @var $textnode \DOMNode */
-      $textnode = $textnodeWrapper->getNode();
-      $xp = $textnode->getNodePath();
-
-      $doSkip = false;
-      foreach (self::$skipTagsForRemoveWhitespace as $pattern) {
-        if (\strpos($xp, "/$pattern") !== false) {
-          $doSkip = true;
-          break;
+        if ($this->doOptimizeViaHtmlDomParser) {
+            $html = $this->minifyHtmlDom($html, $multiDecodeNewHtmlEntity);
         }
-      }
-      if ($doSkip) {
-        continue;
-      }
 
-      $textnode->nodeValue = \preg_replace(self::$regExSpace, ' ', $textnode->nodeValue);
+        // -------------------------------------------------------------------------
+        // Trim whitespace from html-string. [protected html is still protected]
+        // -------------------------------------------------------------------------
+
+        // Remove extra white-space(s) between HTML attribute(s)
+        if (\strpos($html, ' ') !== false) {
+            $html = (string) \preg_replace_callback(
+                '#<([^/\s<>!]+)(?:\s+([^<>]*?)\s*|\s*)(/?)>#',
+                static function ($matches) {
+                    return '<' . $matches[1] . \preg_replace('#([^\s=]+)(=([\'"]?)(.*?)\3)?(\s+|$)#su', ' $1$2', $matches[2]) . $matches[3] . '>';
+                },
+                $html
+            );
+        }
+
+        if ($this->doRemoveSpacesBetweenTags) {
+            /** @noinspection NestedPositiveIfStatementsInspection */
+            if (\strpos($html, ' ') !== false) {
+                // Remove spaces that are between > and <
+                $html = (string) \preg_replace('#(>)\s(<)#', '>$2', $html);
+            }
+        }
+
+        // -------------------------------------------------------------------------
+        // Restore protected HTML-code.
+        // -------------------------------------------------------------------------
+
+        if (\strpos($html, $this->protectedChildNodesHelper) !== false) {
+            $html = (string) \preg_replace_callback(
+                '/<(?<element>' . $this->protectedChildNodesHelper . ')(?<attributes> [^>]*)?>(?<value>.*?)<\/' . $this->protectedChildNodesHelper . '>/',
+                [$this, 'restoreProtectedHtml'],
+                $html
+            );
+        }
+
+        // -------------------------------------------------------------------------
+        // Restore protected HTML-entities.
+        // -------------------------------------------------------------------------
+
+        if ($this->doOptimizeViaHtmlDomParser) {
+            $html = HtmlDomParser::putReplacedBackToPreserveHtmlEntities($html);
+        }
+
+        // ------------------------------------
+        // Final clean-up
+        // ------------------------------------
+
+        $html = \str_replace(
+            [
+                'html>' . "\n",
+                "\n" . '<html',
+                'html/>' . "\n",
+                "\n" . '</html',
+                'head>' . "\n",
+                "\n" . '<head',
+                'head/>' . "\n",
+                "\n" . '</head',
+            ],
+            [
+                'html>',
+                '<html',
+                'html/>',
+                '</html',
+                'head>',
+                '<head',
+                'head/>',
+                '</head',
+            ],
+            $html
+        );
+
+        // self closing tags, don't need a trailing slash ...
+        $replace = [];
+        $replacement = [];
+        foreach (self::$selfClosingTags as $selfClosingTag) {
+            $replace[] = '<' . $selfClosingTag . '/>';
+            $replacement[] = '<' . $selfClosingTag . '>';
+            $replace[] = '<' . $selfClosingTag . ' />';
+            $replacement[] = '<' . $selfClosingTag . '>';
+            $replace[] = '></' . $selfClosingTag . '>';
+            $replacement[] = '>';
+        }
+        $html = \str_replace(
+            $replace,
+            $replacement,
+            $html
+        );
+
+        // ------------------------------------
+        // check if compression worked
+        // ------------------------------------
+
+        if ($origHtmlLength < \strlen($html)) {
+            $html = $origHtml;
+        }
+
+        return $html;
     }
 
-    $dom->getDocument()->normalizeDocument();
+    /**
+     * @param \DOMNode $node
+     *
+     * @return \DOMNode|null
+     */
+    protected function getNextSiblingOfTypeDOMElement(\DOMNode $node)
+    {
+        do {
+            /** @var \DOMNode|null $node - false-positive error from phpstan */
+            $node = $node->nextSibling;
+        } while (!($node === null || $node instanceof \DOMElement));
 
-    return $dom;
-  }
+        return $node;
+    }
+
+    /**
+     * Check if the current string is an conditional comment.
+     *
+     * INFO: since IE >= 10 conditional comment are not working anymore
+     *
+     * <!--[if expression]> HTML <![endif]-->
+     * <![if expression]> HTML <![endif]>
+     *
+     * @param string $comment
+     *
+     * @return bool
+     */
+    private function isConditionalComment($comment): bool
+    {
+        if (\strpos($comment, '[if ') !== false) {
+            /** @noinspection RegExpRedundantEscape */
+            if (\preg_match('/^\[if [^\]]+\]/', $comment)) {
+                return true;
+            }
+        }
+
+        if (\strpos($comment, '[endif]') !== false) {
+            /** @noinspection RegExpRedundantEscape */
+            if (\preg_match('/\[endif\]$/', $comment)) {
+                return true;
+            }
+        }
+
+        return false;
+    }
+
+    /**
+     * @param string $html
+     * @param bool   $multiDecodeNewHtmlEntity
+     *
+     * @return string
+     */
+    private function minifyHtmlDom($html, $multiDecodeNewHtmlEntity): string
+    {
+        // init dom
+        $dom = new HtmlDomParser();
+        /** @noinspection UnusedFunctionResultInspection */
+        $dom->useKeepBrokenHtml($this->keepBrokenHtml);
+
+        $dom->getDocument()->preserveWhiteSpace = false; // remove redundant white space
+        $dom->getDocument()->formatOutput = false; // do not formats output with indentation
+
+        // load dom
+        /** @noinspection UnusedFunctionResultInspection */
+        $dom->loadHtml($html);
+
+        $this->withDocType = (\stripos(\ltrim($html), '<!DOCTYPE') === 0);
+
+        // -------------------------------------------------------------------------
+        // Protect <nocompress> HTML tags first.
+        // -------------------------------------------------------------------------
+
+        $dom = $this->protectTagHelper($dom, 'nocompress');
+
+        // -------------------------------------------------------------------------
+        // Notify the Observer before the minification.
+        // -------------------------------------------------------------------------
+
+        foreach ($dom->find('*') as $element) {
+            $this->notifyObserversAboutDomElementBeforeMinification($element);
+        }
+
+        // -------------------------------------------------------------------------
+        // Protect HTML tags and conditional comments.
+        // -------------------------------------------------------------------------
+
+        $dom = $this->protectTags($dom);
+
+        // -------------------------------------------------------------------------
+        // Remove default HTML comments. [protected html is still protected]
+        // -------------------------------------------------------------------------
+
+        if ($this->doRemoveComments) {
+            $dom = $this->removeComments($dom);
+        }
+
+        // -------------------------------------------------------------------------
+        // Sum-Up extra whitespace from the Dom. [protected html is still protected]
+        // -------------------------------------------------------------------------
+
+        if ($this->doSumUpWhitespace) {
+            $dom = $this->sumUpWhitespace($dom);
+        }
+
+        foreach ($dom->find('*') as $element) {
+
+            // -------------------------------------------------------------------------
+            // Remove whitespace around tags. [protected html is still protected]
+            // -------------------------------------------------------------------------
+
+            if ($this->doRemoveWhitespaceAroundTags) {
+                $this->removeWhitespaceAroundTags($element);
+            }
+
+            // -------------------------------------------------------------------------
+            // Notify the Observer after the minification.
+            // -------------------------------------------------------------------------
+
+            $this->notifyObserversAboutDomElementAfterMinification($element);
+        }
+
+        // -------------------------------------------------------------------------
+        // Convert the Dom into a string.
+        // -------------------------------------------------------------------------
+
+        return $dom->fixHtmlOutput(
+            $this->domNodeToString($dom->getDocument()),
+            $multiDecodeNewHtmlEntity
+        );
+    }
+
+    /**
+     * @param SimpleHtmlDomInterface $domElement
+     *
+     * @return void
+     */
+    private function notifyObserversAboutDomElementAfterMinification(SimpleHtmlDomInterface $domElement)
+    {
+        foreach ($this->domLoopObservers as $observer) {
+            $observer->domElementAfterMinification($domElement, $this);
+        }
+    }
+
+    /**
+     * @param SimpleHtmlDomInterface $domElement
+     *
+     * @return void
+     */
+    private function notifyObserversAboutDomElementBeforeMinification(SimpleHtmlDomInterface $domElement)
+    {
+        foreach ($this->domLoopObservers as $observer) {
+            $observer->domElementBeforeMinification($domElement, $this);
+        }
+    }
+
+    /**
+     * @param HtmlDomParser $dom
+     * @param string        $selector
+     *
+     * @return HtmlDomParser
+     */
+    private function protectTagHelper(HtmlDomParser $dom, string $selector): HtmlDomParser
+    {
+        foreach ($dom->find($selector) as $element) {
+            if ($element->isRemoved()) {
+                continue;
+            }
+
+            $this->protectedChildNodes[$this->protected_tags_counter] = $element->parentNode()->innerHtml();
+            $element->getNode()->parentNode->nodeValue = '<' . $this->protectedChildNodesHelper . ' data-' . $this->protectedChildNodesHelper . '="' . $this->protected_tags_counter . '"></' . $this->protectedChildNodesHelper . '>';
+
+            ++$this->protected_tags_counter;
+        }
+
+        return $dom;
+    }
+
+    /**
+     * Prevent changes of inline "styles" and "scripts".
+     *
+     * @param HtmlDomParser $dom
+     *
+     * @return HtmlDomParser
+     */
+    private function protectTags(HtmlDomParser $dom): HtmlDomParser
+    {
+        $this->protectTagHelper($dom, 'code');
+
+        foreach ($dom->find('script, style') as $element) {
+            if ($element->isRemoved()) {
+                continue;
+            }
+
+            if ($element->tag === 'script' || $element->tag === 'style') {
+                $attributes = $element->getAllAttributes();
+                // skip external links
+                if (isset($attributes['src'])) {
+                    continue;
+                }
+            }
+
+            $this->protectedChildNodes[$this->protected_tags_counter] = $element->innerhtml;
+            $element->getNode()->nodeValue = '<' . $this->protectedChildNodesHelper . ' data-' . $this->protectedChildNodesHelper . '="' . $this->protected_tags_counter . '"></' . $this->protectedChildNodesHelper . '>';
+
+            ++$this->protected_tags_counter;
+        }
+
+        foreach ($dom->find('//comment()') as $element) {
+            if ($element->isRemoved()) {
+                continue;
+            }
+
+            $text = $element->text();
+
+            // skip normal comments
+            if (!$this->isConditionalComment($text)) {
+                continue;
+            }
+
+            $this->protectedChildNodes[$this->protected_tags_counter] = '<!--' . $text . '-->';
+
+            /* @var $node \DOMComment */
+            $node = $element->getNode();
+            $child = new \DOMText('<' . $this->protectedChildNodesHelper . ' data-' . $this->protectedChildNodesHelper . '="' . $this->protected_tags_counter . '"></' . $this->protectedChildNodesHelper . '>');
+            /** @noinspection UnusedFunctionResultInspection */
+            $element->getNode()->parentNode->replaceChild($child, $node);
+
+            ++$this->protected_tags_counter;
+        }
+
+        return $dom;
+    }
+
+    /**
+     * Remove comments in the dom.
+     *
+     * @param HtmlDomParser $dom
+     *
+     * @return HtmlDomParser
+     */
+    private function removeComments(HtmlDomParser $dom): HtmlDomParser
+    {
+        foreach ($dom->find('//comment()') as $commentWrapper) {
+            $comment = $commentWrapper->getNode();
+            $val = $comment->nodeValue;
+            if (\strpos($val, '[') === false) {
+                /** @noinspection UnusedFunctionResultInspection */
+                $comment->parentNode->removeChild($comment);
+            }
+        }
+
+        $dom->getDocument()->normalizeDocument();
+
+        return $dom;
+    }
+
+    /**
+     * Trim tags in the dom.
+     *
+     * @param SimpleHtmlDomInterface $element
+     *
+     * @return void
+     */
+    private function removeWhitespaceAroundTags(SimpleHtmlDomInterface $element)
+    {
+        if (isset(self::$trimWhitespaceFromTags[$element->tag])) {
+            $node = $element->getNode();
+
+            /** @var \DOMNode[] $candidates */
+            $candidates = [];
+            if ($node->childNodes->length > 0) {
+                $candidates[] = $node->firstChild;
+                $candidates[] = $node->lastChild;
+                $candidates[] = $node->previousSibling;
+                $candidates[] = $node->nextSibling;
+            }
+
+            /** @var mixed $candidate - false-positive error from phpstan */
+            foreach ($candidates as &$candidate) {
+                if ($candidate === null) {
+                    continue;
+                }
+
+                if ($candidate->nodeType === \XML_TEXT_NODE) {
+                    $nodeValueTmp = \preg_replace(self::$regExSpace, ' ', $candidate->nodeValue);
+                    if ($nodeValueTmp !== null) {
+                        $candidate->nodeValue = $nodeValueTmp;
+                    }
+                }
+            }
+        }
+    }
+
+    /**
+     * Callback function for preg_replace_callback use.
+     *
+     * @param array $matches PREG matches
+     *
+     * @return string
+     */
+    private function restoreProtectedHtml($matches): string
+    {
+        \preg_match('/.*"(?<id>\d*)"/', $matches['attributes'], $matchesInner);
+
+        return $this->protectedChildNodes[$matchesInner['id']] ?? '';
+    }
+
+    /**
+     * @param array $domainsToRemoveHttpPrefixFromAttributes
+     *
+     * @return $this
+     */
+    public function setDomainsToRemoveHttpPrefixFromAttributes($domainsToRemoveHttpPrefixFromAttributes): self
+    {
+        $this->domainsToRemoveHttpPrefixFromAttributes = $domainsToRemoveHttpPrefixFromAttributes;
+
+        return $this;
+    }
+
+    /**
+     * Sum-up extra whitespace from dom-nodes.
+     *
+     * @param HtmlDomParser $dom
+     *
+     * @return HtmlDomParser
+     */
+    private function sumUpWhitespace(HtmlDomParser $dom): HtmlDomParser
+    {
+        $text_nodes = $dom->find('//text()');
+        foreach ($text_nodes as $text_node_wrapper) {
+            /* @var $text_node \DOMNode */
+            $text_node = $text_node_wrapper->getNode();
+            $xp = $text_node->getNodePath();
+            if ($xp === null) {
+                continue;
+            }
+
+            $doSkip = false;
+            foreach (self::$skipTagsForRemoveWhitespace as $pattern) {
+                if (\strpos($xp, "/${pattern}") !== false) {
+                    $doSkip = true;
+
+                    break;
+                }
+            }
+            if ($doSkip) {
+                continue;
+            }
+
+            $nodeValueTmp = \preg_replace(self::$regExSpace, ' ', $text_node->nodeValue);
+            if ($nodeValueTmp !== null) {
+                $text_node->nodeValue = $nodeValueTmp;
+            }
+        }
+
+        $dom->getDocument()->normalizeDocument();
+
+        return $dom;
+    }
+
+    /**
+     * WARNING: maybe bad for performance ...
+     *
+     * @param bool $keepBrokenHtml
+     *
+     * @return HtmlMin
+     */
+    public function useKeepBrokenHtml(bool $keepBrokenHtml): self
+    {
+        $this->keepBrokenHtml = $keepBrokenHtml;
+
+        return $this;
+    }
 }
