@@ -4,731 +4,914 @@ declare(strict_types=1);
 
 namespace voku\helper;
 
-use BadMethodCallException;
-use DOMElement;
-use DOMNode;
-use RuntimeException;
-
 /**
- * Class SimpleHtmlDom
+ * @noinspection PhpHierarchyChecksInspection
  *
- * @package voku\helper
+ * {@inheritdoc}
  *
- * @property string      outerText <p>Get dom node's outer html (alias for "outerHtml").</p>
- * @property string      outerHtml <p>Get dom node's outer html.</p>
- * @property string      innerText <p>Get dom node's inner html (alias for "innerHtml").</p>
- * @property string      innerHtml <p>Get dom node's inner html.</p>
- * @property-read string plaintext <p>Get dom node's plain text.</p>
- * @property-read string tag       <p>Get dom node name.</p>
- * @property-read string attr      <p>Get dom node attributes.</p>
- *
- * @method SimpleHtmlDomNode|SimpleHtmlDom|null children() children($idx = -1) <p>Returns children of node.</p>
- * @method SimpleHtmlDom|null first_child() <p>Returns the first child of node.</p>
- * @method SimpleHtmlDom|null last_child() <p>Returns the last child of node.</p>
- * @method SimpleHtmlDom|null next_sibling() <p>Returns the next sibling of node.</p>
- * @method SimpleHtmlDom|null prev_sibling() <p>Returns the previous sibling of node.</p>
- * @method SimpleHtmlDom|null parent() <p>Returns the parent of node.</p>
- *
- * @method string outerText() <p>Get dom node's outer html (alias for "outerHtml()").</p>
- * @method string outerHtml() <p>Get dom node's outer html.</p>
- * @method string innerText() <p>Get dom node's inner html (alias for "innerHtml()").</p>
- *
+ * @implements \IteratorAggregate<int, \DOMNode>
  */
-class SimpleHtmlDom implements \IteratorAggregate
+class SimpleHtmlDom extends AbstractSimpleHtmlDom implements \IteratorAggregate, SimpleHtmlDomInterface
 {
-  /**
-   * @var array
-   */
-  protected static $functionAliases = array(
-      'children'     => 'childNodes',
-      'first_child'  => 'firstChild',
-      'last_child'   => 'lastChild',
-      'next_sibling' => 'nextSibling',
-      'prev_sibling' => 'previousSibling',
-      'parent'       => 'parentNode',
-      'outertext'    => 'html',
-      'outerhtml'    => 'html',
-      'innertext'    => 'innerHtml',
-      'innerhtml'    => 'innerHtml',
-  );
-
-  /**
-   * @var DOMElement
-   */
-  protected $node;
-
-  /**
-   * SimpleHtmlDom constructor.
-   *
-   * @param DOMNode $node
-   */
-  public function __construct(DOMNode $node)
-  {
-    $this->node = $node;
-  }
-
-  /**
-   * @param string $name
-   * @param array $arguments
-   *
-   * @return null|string|SimpleHtmlDom
-   *
-   * @throws \BadMethodCallException
-   */
-  public function __call($name, $arguments)
-  {
-    $name = \strtolower($name);
-
-    if (isset(self::$functionAliases[$name])) {
-      return \call_user_func_array(array($this, self::$functionAliases[$name]), $arguments);
+    /**
+     * @param \DOMElement|\DOMNode $node
+     */
+    public function __construct(\DOMNode $node)
+    {
+        $this->node = $node;
     }
 
-    throw new BadMethodCallException('Method does not exist');
-  }
+    /**
+     * @param string $name
+     * @param array  $arguments
+     *
+     * @throws \BadMethodCallException
+     *
+     * @return SimpleHtmlDomInterface|string|null
+     */
+    public function __call($name, $arguments)
+    {
+        $name = \strtolower($name);
 
-  /**
-   * @param string $name
-   *
-   * @return array|null|string
-   */
-  public function __get($name)
-  {
-    $name = \strtolower($name);
+        if (isset(self::$functionAliases[$name])) {
+            return \call_user_func_array([$this, self::$functionAliases[$name]], $arguments);
+        }
 
-    switch ($name) {
-      case 'outerhtml':
-      case 'outertext':
-        return $this->html();
-      case 'innerhtml':
-      case 'innertext':
-        return $this->innerHtml();
-      case 'text':
-      case 'plaintext':
-        return $this->text();
-      case 'tag':
-        return $this->node->nodeName;
-      case 'attr':
-        return $this->getAllAttributes();
-      default:
-        return $this->getAttribute($name);
-    }
-  }
-
-  /**
-   * @param string $selector
-   * @param int    $idx
-   *
-   * @return SimpleHtmlDom[]|SimpleHtmlDom|SimpleHtmlDomNodeInterface
-   */
-  public function __invoke($selector, $idx = null)
-  {
-    return $this->find($selector, $idx);
-  }
-
-  /**
-   * @param $name
-   *
-   * @return bool
-   */
-  public function __isset($name)
-  {
-    $name = strtolower($name);
-
-    switch ($name) {
-      case 'outertext':
-      case 'outerhtml':
-      case 'innertext':
-      case 'innerhtml':
-      case 'plaintext':
-      case 'text':
-      case 'tag':
-        return true;
-      default:
-        return $this->hasAttribute($name);
-    }
-  }
-
-  /**
-   * @param $name
-   * @param $value
-   *
-   * @return SimpleHtmlDom
-   */
-  public function __set($name, $value)
-  {
-    $name = strtolower($name);
-
-    switch ($name) {
-      case 'outerhtml':
-      case 'outertext':
-        return $this->replaceNode($value);
-      case 'innertext':
-      case 'innerhtml':
-        return $this->replaceChild($value);
-      default:
-        return $this->setAttribute($name, $value);
-    }
-  }
-
-  /**
-   * @return string
-   */
-  public function __toString()
-  {
-    return $this->html();
-  }
-
-  /**
-   * @param $name
-   *
-   * @return SimpleHtmlDom
-   */
-  public function __unset($name)
-  {
-    return $this->removeAttribute($name);
-  }
-
-  /**
-   * Returns children of node.
-   *
-   * @param int $idx
-   *
-   * @return SimpleHtmlDomNode|SimpleHtmlDom|null
-   */
-  public function childNodes(int $idx = -1)
-  {
-    $nodeList = $this->getIterator();
-
-    if ($idx === -1) {
-      return $nodeList;
+        throw new \BadMethodCallException('Method does not exist');
     }
 
-    if (isset($nodeList[$idx])) {
-      return $nodeList[$idx];
+    /**
+     * Find list of nodes with a CSS selector.
+     *
+     * @param string   $selector
+     * @param int|null $idx
+     *
+     * @return SimpleHtmlDomInterface|SimpleHtmlDomInterface[]|SimpleHtmlDomNodeInterface<SimpleHtmlDomInterface>
+     */
+    public function find(string $selector, $idx = null)
+    {
+        return $this->getHtmlDomParser()->find($selector, $idx);
     }
 
-    return null;
-  }
+    /**
+     * Returns an array of attributes.
+     *
+     * @return string[]|null
+     */
+    public function getAllAttributes()
+    {
+        if ($this->node->hasAttributes()) {
+            $attributes = [];
+            foreach ($this->node->attributes as $attr) {
+                $attributes[$attr->name] = HtmlDomParser::putReplacedBackToPreserveHtmlEntities($attr->value);
+            }
 
-  /**
-   * Find list of nodes with a CSS selector.
-   *
-   * @param string   $selector
-   * @param int|null $idx
-   *
-   * @return SimpleHtmlDom[]|SimpleHtmlDom|SimpleHtmlDomNodeInterface
-   */
-  public function find(string $selector, $idx = null)
-  {
-    return $this->getHtmlDomParser()->find($selector, $idx);
-  }
+            return $attributes;
+        }
 
-  /**
-   * Find one node with a CSS selector.
-   *
-   * @param string $selector
-   *
-   * @return SimpleHtmlDom|SimpleHtmlDomNodeInterface
-   */
-  public function findOne(string $selector)
-  {
-    return $this->find($selector, 0);
-  }
-
-  /**
-   * Returns the first child of node.
-   *
-   * @return SimpleHtmlDom|null
-   */
-  public function firstChild()
-  {
-    $node = $this->node->firstChild;
-
-    if ($node === null) {
-      return null;
+        return null;
     }
 
-    return new self($node);
-  }
-
-  /**
-   * Returns an array of attributes.
-   *
-   * @return array|null
-   */
-  public function getAllAttributes()
-  {
-    if ($this->node->hasAttributes()) {
-      $attributes = array();
-      foreach ($this->node->attributes as $attr) {
-        $attributes[$attr->name] = HtmlDomParser::putReplacedBackToPreserveHtmlEntities($attr->value);
-      }
-
-      return $attributes;
+    /**
+     * @return bool
+     */
+    public function hasAttributes(): bool
+    {
+        return $this->node->hasAttributes();
     }
 
-    return null;
-  }
+    /**
+     * Return attribute value.
+     *
+     * @param string $name
+     *
+     * @return string
+     */
+    public function getAttribute(string $name): string
+    {
+        if ($this->node instanceof \DOMElement) {
+            return HtmlDomParser::putReplacedBackToPreserveHtmlEntities(
+                $this->node->getAttribute($name)
+            );
+        }
 
-  /**
-   * Return attribute value.
-   *
-   * @param string $name
-   *
-   * @return string
-   */
-  public function getAttribute(string $name): string
-  {
-    $html = $this->node->getAttribute($name);
-
-    return HtmlDomParser::putReplacedBackToPreserveHtmlEntities($html);
-  }
-
-  /**
-   * Return element by #id.
-   *
-   * @param string $id
-   *
-   * @return SimpleHtmlDom|SimpleHtmlDomNodeInterface
-   */
-  public function getElementById(string $id)
-  {
-    return $this->find("#$id", 0);
-  }
-
-  /**
-   * Return element by tag name.
-   *
-   * @param string $name
-   *
-   * @return SimpleHtmlDom|SimpleHtmlDomNodeBlank
-   */
-  public function getElementByTagName(string $name)
-  {
-    $node = $this->node->getElementsByTagName($name)->item(0);
-
-    if ($node === null) {
-      return new SimpleHtmlDomNodeBlank();
+        return '';
     }
 
-    return new self($node);
-  }
+    /**
+     * Determine if an attribute exists on the element.
+     *
+     * @param string $name
+     *
+     * @return bool
+     */
+    public function hasAttribute(string $name): bool
+    {
+        if (!$this->node instanceof \DOMElement) {
+            return false;
+        }
 
-  /**
-   * Returns elements by #id.
-   *
-   * @param string   $id
-   * @param null|int $idx
-   *
-   * @return SimpleHtmlDom|SimpleHtmlDom[]|SimpleHtmlDomNodeInterface
-   */
-  public function getElementsById(string $id, $idx = null)
-  {
-    return $this->find("#$id", $idx);
-  }
-
-  /**
-   * Returns elements by tag name.
-   *
-   * @param string   $name
-   * @param null|int $idx
-   *
-   * @return SimpleHtmlDomNode|SimpleHtmlDom[]|SimpleHtmlDomNodeBlank
-   */
-  public function getElementsByTagName(string $name, $idx = null)
-  {
-    $nodesList = $this->node->getElementsByTagName($name);
-
-    $elements = new SimpleHtmlDomNode();
-
-    foreach ($nodesList as $node) {
-      $elements[] = new self($node);
+        return $this->node->hasAttribute($name);
     }
 
-    // return all elements
-    if (null === $idx) {
-      return $elements;
+    /**
+     * Get dom node's outer html.
+     *
+     * @param bool $multiDecodeNewHtmlEntity
+     *
+     * @return string
+     */
+    public function html(bool $multiDecodeNewHtmlEntity = false): string
+    {
+        return $this->getHtmlDomParser()->html($multiDecodeNewHtmlEntity);
     }
 
-    // handle negative values
-    if ($idx < 0) {
-      $idx = \count($elements) + $idx;
+    /**
+     * Get dom node's inner html.
+     *
+     * @param bool $multiDecodeNewHtmlEntity
+     *
+     * @return string
+     */
+    public function innerHtml(bool $multiDecodeNewHtmlEntity = false): string
+    {
+        return $this->getHtmlDomParser()->innerHtml($multiDecodeNewHtmlEntity);
     }
 
-    // return one element
-    if (isset($elements[$idx])) {
-      return $elements[$idx];
+    /**
+     * Remove attribute.
+     *
+     * @param string $name <p>The name of the html-attribute.</p>
+     *
+     * @return SimpleHtmlDomInterface
+     */
+    public function removeAttribute(string $name): SimpleHtmlDomInterface
+    {
+        if (\method_exists($this->node, 'removeAttribute')) {
+            $this->node->removeAttribute($name);
+        }
+
+        return $this;
     }
 
-    // return a blank-element
-    return new SimpleHtmlDomNodeBlank();
-  }
+    /**
+     * Replace child node.
+     *
+     * @param string $string
+     *
+     * @return SimpleHtmlDomInterface
+     */
+    protected function replaceChildWithString(string $string): SimpleHtmlDomInterface
+    {
+        if (!empty($string)) {
+            $newDocument = new HtmlDomParser($string);
 
-  /**
-   * Create a new "HtmlDomParser"-object from the current context.
-   *
-   * @return HtmlDomParser
-   */
-  public function getHtmlDomParser(): HtmlDomParser
-  {
-    return new HtmlDomParser($this);
-  }
+            $tmpDomString = $this->normalizeStringForComparision($newDocument);
+            $tmpStr = $this->normalizeStringForComparision($string);
+            if ($tmpDomString !== $tmpStr) {
+                throw new \RuntimeException(
+                    'Not valid HTML fragment!' . "\n" .
+                    $tmpDomString . "\n" .
+                    $tmpStr
+                );
+            }
+        }
 
-  /**
-   * Retrieve an external iterator.
-   *
-   * @link  http://php.net/manual/en/iteratoraggregate.getiterator.php
-   * @return SimpleHtmlDomNode An instance of an object implementing <b>Iterator</b> or
-   * <b>Traversable</b>
-   */
-  public function getIterator(): SimpleHtmlDomNode
-  {
-    $elements = new SimpleHtmlDomNode();
-    if ($this->node->hasChildNodes()) {
-      foreach ($this->node->childNodes as $node) {
-        $elements[] = new self($node);
-      }
+        /** @var \DOMNode[] $remove_nodes */
+        $remove_nodes = [];
+        if ($this->node->childNodes->length > 0) {
+            // INFO: We need to fetch the nodes first, before we can delete them, because of missing references in the dom,
+            // if we delete the elements on the fly.
+            foreach ($this->node->childNodes as $node) {
+                $remove_nodes[] = $node;
+            }
+        }
+        foreach ($remove_nodes as $remove_node) {
+            $this->node->removeChild($remove_node);
+        }
+
+        if (!empty($newDocument)) {
+            $newDocument = $this->cleanHtmlWrapper($newDocument);
+            $ownerDocument = $this->node->ownerDocument;
+            if (
+                $ownerDocument !== null
+                &&
+                $newDocument->getDocument()->documentElement !== null
+            ) {
+                $newNode = $ownerDocument->importNode($newDocument->getDocument()->documentElement, true);
+                /** @noinspection UnusedFunctionResultInspection */
+                $this->node->appendChild($newNode);
+            }
+        }
+
+        return $this;
     }
 
-    return $elements;
-  }
+    /**
+     * Replace this node.
+     *
+     * @param string $string
+     *
+     * @return SimpleHtmlDomInterface
+     */
+    protected function replaceNodeWithString(string $string): SimpleHtmlDomInterface
+    {
+        if (empty($string)) {
+            $this->node->parentNode->removeChild($this->node);
 
-  /**
-   * @return DOMNode
-   */
-  public function getNode(): \DOMNode
-  {
-    return $this->node;
-  }
+            return $this;
+        }
 
-  /**
-   * Determine if an attribute exists on the element.
-   *
-   * @param string $name
-   *
-   * @return bool
-   */
-  public function hasAttribute(string $name): bool
-  {
-    return $this->node->hasAttribute($name);
-  }
+        $newDocument = new HtmlDomParser($string);
 
-  /**
-   * Get dom node's outer html.
-   *
-   * @param bool $multiDecodeNewHtmlEntity
-   *
-   * @return string
-   */
-  public function html(bool $multiDecodeNewHtmlEntity = false): string
-  {
-    return $this->getHtmlDomParser()->html($multiDecodeNewHtmlEntity);
-  }
+        $tmpDomOuterTextString = $this->normalizeStringForComparision($newDocument);
+        $tmpStr = $this->normalizeStringForComparision($string);
+        if ($tmpDomOuterTextString !== $tmpStr) {
+            throw new \RuntimeException(
+                'Not valid HTML fragment!' . "\n"
+                . $tmpDomOuterTextString . "\n" .
+                $tmpStr
+            );
+        }
 
-  /**
-   * Get dom node's inner html.
-   *
-   * @param bool $multiDecodeNewHtmlEntity
-   *
-   * @return string
-   */
-  public function innerHtml(bool $multiDecodeNewHtmlEntity = false): string
-  {
-    return $this->getHtmlDomParser()->innerHtml($multiDecodeNewHtmlEntity);
-  }
+        $newDocument = $this->cleanHtmlWrapper($newDocument, true);
+        $ownerDocument = $this->node->ownerDocument;
+        if (
+            $ownerDocument === null
+            ||
+            $newDocument->getDocument()->documentElement === null
+        ) {
+            return $this;
+        }
 
-  /**
-   * Returns the last child of node.
-   *
-   * @return SimpleHtmlDom|null
-   */
-  public function lastChild()
-  {
-    $node = $this->node->lastChild;
+        $newNode = $ownerDocument->importNode($newDocument->getDocument()->documentElement, true);
 
-    if ($node === null) {
-      return null;
+        $this->node->parentNode->replaceChild($newNode, $this->node);
+        $this->node = $newNode;
+
+        // Remove head element, preserving child nodes. (again)
+        if (
+            $this->node->parentNode instanceof \DOMElement
+            &&
+            $newDocument->getIsDOMDocumentCreatedWithoutHeadWrapper()
+        ) {
+            $html = $this->node->parentNode->getElementsByTagName('head')[0];
+
+            if ($this->node->parentNode->ownerDocument !== null) {
+                $fragment = $this->node->parentNode->ownerDocument->createDocumentFragment();
+                if ($html !== null) {
+                    /** @var \DOMNode $html */
+                    while ($html->childNodes->length > 0) {
+                        $tmpNode = $html->childNodes->item(0);
+                        if ($tmpNode !== null) {
+                            /** @noinspection UnusedFunctionResultInspection */
+                            $fragment->appendChild($tmpNode);
+                        }
+                    }
+                    /** @noinspection UnusedFunctionResultInspection */
+                    $html->parentNode->replaceChild($fragment, $html);
+                }
+            }
+        }
+
+        return $this;
     }
 
-    return new self($node);
-  }
+    /**
+     * Replace this node with text
+     *
+     * @param string $string
+     *
+     * @return SimpleHtmlDomInterface
+     */
+    protected function replaceTextWithString($string): SimpleHtmlDomInterface
+    {
+        if (empty($string)) {
+            $this->node->parentNode->removeChild($this->node);
 
-  /**
-   * Returns the next sibling of node.
-   *
-   * @return SimpleHtmlDom|null
-   */
-  public function nextSibling()
-  {
-    $node = $this->node->nextSibling;
+            return $this;
+        }
 
-    if ($node === null) {
-      return null;
+        $ownerDocument = $this->node->ownerDocument;
+        if ($ownerDocument !== null) {
+            $newElement = $ownerDocument->createTextNode($string);
+            $newNode = $ownerDocument->importNode($newElement, true);
+            $this->node->parentNode->replaceChild($newNode, $this->node);
+            $this->node = $newNode;
+        }
+
+        return $this;
     }
 
-    return new self($node);
-  }
+    /**
+     * Set attribute value.
+     *
+     * @param string      $name       <p>The name of the html-attribute.</p>
+     * @param string|null $value      <p>Set to NULL or empty string, to remove the attribute.</p>
+     * @param bool        $strict     </p>
+     *                                $value must be NULL, to remove the attribute,
+     *                                so that you can set an empty string as attribute-value e.g. autofocus=""
+     *                                </p>
+     *
+     * @return SimpleHtmlDomInterface
+     */
+    public function setAttribute(string $name, $value = null, bool $strict = false): SimpleHtmlDomInterface
+    {
+        if (
+            ($strict && $value === null)
+            ||
+            (!$strict && empty($value))
+        ) {
+            /** @noinspection UnusedFunctionResultInspection */
+            $this->removeAttribute($name);
+        } elseif (\method_exists($this->node, 'setAttribute')) {
+            /** @noinspection UnusedFunctionResultInspection */
+            $this->node->setAttribute($name, $value);
+        }
 
-  /**
-   * Returns the parent of node.
-   *
-   * @return SimpleHtmlDom
-   */
-  public function parentNode(): self
-  {
-    return new self($this->node->parentNode);
-  }
-
-  /**
-   * Returns the previous sibling of node.
-   *
-   * @return SimpleHtmlDom|null
-   */
-  public function previousSibling()
-  {
-    $node = $this->node->previousSibling;
-
-    if ($node === null) {
-      return null;
+        return $this;
     }
 
-    return new self($node);
-  }
-
-  /**
-   * Replace child node.
-   *
-   * @param string $string
-   *
-   * @return $this
-   *
-   * @throws \RuntimeException
-   */
-  protected function replaceChild(string $string)
-  {
-    if (!empty($string)) {
-      $newDocument = new HtmlDomParser($string);
-
-      if ($this->normalizeStringForComparision($newDocument) != $this->normalizeStringForComparision($string)) {
-        throw new RuntimeException('Not valid HTML fragment');
-      }
+    /**
+     * Get dom node's plain text.
+     *
+     * @return string
+     */
+    public function text(): string
+    {
+        return $this->getHtmlDomParser()->fixHtmlOutput($this->node->textContent);
     }
 
-    /** @noinspection PhpParamsInspection */
-    if (\count($this->node->childNodes) > 0) {
-      foreach ($this->node->childNodes as $node) {
-        $this->node->removeChild($node);
-      }
+    /**
+     * Change the name of a tag in a "DOMNode".
+     *
+     * @param \DOMNode $node
+     * @param string   $name
+     *
+     * @return \DOMElement|false
+     *                          <p>DOMElement a new instance of class DOMElement or false
+     *                          if an error occured.</p>
+     */
+    protected function changeElementName(\DOMNode $node, string $name)
+    {
+        $ownerDocument = $node->ownerDocument;
+        if ($ownerDocument) {
+            $newNode = $ownerDocument->createElement($name);
+        } else {
+            return false;
+        }
+
+        foreach ($node->childNodes as $child) {
+            $child = $ownerDocument->importNode($child, true);
+            /** @noinspection UnusedFunctionResultInspection */
+            $newNode->appendChild($child);
+        }
+
+        foreach ($node->attributes as $attrName => $attrNode) {
+            /** @noinspection UnusedFunctionResultInspection */
+            $newNode->setAttribute($attrName, $attrNode);
+        }
+
+        /** @noinspection UnusedFunctionResultInspection */
+        $newNode->ownerDocument->replaceChild($newNode, $node);
+
+        return $newNode;
     }
 
-    if (!empty($newDocument)) {
-      $newDocument = $this->cleanHtmlWrapper($newDocument);
-      $newNode = $this->node->ownerDocument->importNode($newDocument->getDocument()->documentElement, true);
-      $this->node->appendChild($newNode);
+    /**
+     * Returns children of node.
+     *
+     * @param int $idx
+     *
+     * @return SimpleHtmlDomInterface|SimpleHtmlDomInterface[]|SimpleHtmlDomNodeInterface|null
+     */
+    public function childNodes(int $idx = -1)
+    {
+        $nodeList = $this->getIterator();
+
+        if ($idx === -1) {
+            return $nodeList;
+        }
+
+        return $nodeList[$idx] ?? null;
     }
 
-    return $this;
-  }
-
-  /**
-   * Replace this node.
-   *
-   * @param string $string
-   *
-   * @return $this|null
-   *
-   * @throws \RuntimeException
-   */
-  protected function replaceNode(string $string)
-  {
-    if (empty($string)) {
-      $this->node->parentNode->removeChild($this->node);
-
-      return null;
+    /**
+     * Find nodes with a CSS selector.
+     *
+     * @param string $selector
+     *
+     * @return SimpleHtmlDomInterface[]|SimpleHtmlDomNodeInterface<SimpleHtmlDomInterface>
+     */
+    public function findMulti(string $selector): SimpleHtmlDomNodeInterface
+    {
+        return $this->getHtmlDomParser()->findMulti($selector);
     }
 
-    $newDocument = new HtmlDomParser($string);
-
-    if ($this->normalizeStringForComparision($newDocument->outerText()) != $this->normalizeStringForComparision($string)) {
-      throw new RuntimeException('Not valid HTML fragment');
+    /**
+     * Find nodes with a CSS selector or false, if no element is found.
+     *
+     * @param string $selector
+     *
+     * @return false|SimpleHtmlDomInterface[]|SimpleHtmlDomNodeInterface<SimpleHtmlDomInterface>
+     */
+    public function findMultiOrFalse(string $selector)
+    {
+        return $this->getHtmlDomParser()->findMultiOrFalse($selector);
     }
 
-    $newDocument = $this->cleanHtmlWrapper($newDocument);
-
-    $newNode = $this->node->ownerDocument->importNode($newDocument->getDocument()->documentElement, true);
-
-    $this->node->parentNode->replaceChild($newNode, $this->node);
-    $this->node = $newNode;
-
-    return $this;
-  }
-
-  /**
-   * Normalize the given input for comparision.
-   *
-   * @param HtmlDomParser|string $input
-   *
-   * @return string
-   */
-  private function normalizeStringForComparision($input): string
-  {
-    if ($input instanceof HtmlDomParser) {
-      $string = $input->outerText();
-
-      if ($input->getIsDOMDocumentCreatedWithoutHeadWrapper() === true) {
-        $string = str_replace(array('<head>', '</head>'), '', $string);
-      }
-    } else {
-      $string = (string)$input;
+    /**
+     * Find one node with a CSS selector.
+     *
+     * @param string $selector
+     *
+     * @return SimpleHtmlDomInterface
+     */
+    public function findOne(string $selector): SimpleHtmlDomInterface
+    {
+        return $this->getHtmlDomParser()->findOne($selector);
     }
 
-    return
-        urlencode(
-            urldecode(
-                trim(
-                    str_replace(
-                        array(
-                            ' ',
-                            "\n",
-                            "\r",
-                            '/>',
-                        ),
-                        array(
-                            '',
-                            '',
-                            '',
-                            '>',
-                        ),
-                        strtolower($string)
+    /**
+     * Find one node with a CSS selector or false, if no element is found.
+     *
+     * @param string $selector
+     *
+     * @return false|SimpleHtmlDomInterface
+     */
+    public function findOneOrFalse(string $selector)
+    {
+        return $this->getHtmlDomParser()->findOneOrFalse($selector);
+    }
+
+    /**
+     * Returns the first child of node.
+     *
+     * @return SimpleHtmlDomInterface|null
+     */
+    public function firstChild()
+    {
+        /** @var \DOMNode|null $node */
+        $node = $this->node->firstChild;
+
+        if ($node === null) {
+            return null;
+        }
+
+        return new static($node);
+    }
+
+    /**
+     * Return elements by ".class".
+     *
+     * @param string $class
+     *
+     * @return SimpleHtmlDomInterface[]|SimpleHtmlDomNodeInterface<SimpleHtmlDomInterface>
+     */
+    public function getElementByClass(string $class): SimpleHtmlDomNodeInterface
+    {
+        return $this->findMulti(".${class}");
+    }
+
+    /**
+     * Return element by #id.
+     *
+     * @param string $id
+     *
+     * @return SimpleHtmlDomInterface
+     */
+    public function getElementById(string $id): SimpleHtmlDomInterface
+    {
+        return $this->findOne("#${id}");
+    }
+
+    /**
+     * Return element by tag name.
+     *
+     * @param string $name
+     *
+     * @return SimpleHtmlDomInterface
+     */
+    public function getElementByTagName(string $name): SimpleHtmlDomInterface
+    {
+        if ($this->node instanceof \DOMElement) {
+            $node = $this->node->getElementsByTagName($name)->item(0);
+        } else {
+            $node = null;
+        }
+
+        if ($node === null) {
+            return new SimpleHtmlDomBlank();
+        }
+
+        return new static($node);
+    }
+
+    /**
+     * Returns elements by "#id".
+     *
+     * @param string   $id
+     * @param int|null $idx
+     *
+     * @return SimpleHtmlDomInterface|SimpleHtmlDomInterface[]|SimpleHtmlDomNodeInterface<SimpleHtmlDomInterface>
+     */
+    public function getElementsById(string $id, $idx = null)
+    {
+        return $this->find("#${id}", $idx);
+    }
+
+    /**
+     * Returns elements by tag name.
+     *
+     * @param string   $name
+     * @param int|null $idx
+     *
+     * @return SimpleHtmlDomInterface|SimpleHtmlDomInterface[]|SimpleHtmlDomNodeInterface<SimpleHtmlDomInterface>
+     */
+    public function getElementsByTagName(string $name, $idx = null)
+    {
+        if ($this->node instanceof \DOMElement) {
+            $nodesList = $this->node->getElementsByTagName($name);
+        } else {
+            $nodesList = [];
+        }
+
+        $elements = new SimpleHtmlDomNode();
+
+        foreach ($nodesList as $node) {
+            $elements[] = new static($node);
+        }
+
+        // return all elements
+        if ($idx === null) {
+            if (\count($elements) === 0) {
+                return new SimpleHtmlDomNodeBlank();
+            }
+
+            return $elements;
+        }
+
+        // handle negative values
+        if ($idx < 0) {
+            $idx = \count($elements) + $idx;
+        }
+
+        // return one element
+        return $elements[$idx] ?? new SimpleHtmlDomBlank();
+    }
+
+    /**
+     * Create a new "HtmlDomParser"-object from the current context.
+     *
+     * @return HtmlDomParser
+     */
+    public function getHtmlDomParser(): HtmlDomParser
+    {
+        return new HtmlDomParser($this);
+    }
+
+    /**
+     * @return \DOMNode
+     */
+    public function getNode(): \DOMNode
+    {
+        return $this->node;
+    }
+
+    /**
+     * Nodes can get partially destroyed in which they're still an
+     * actual DOM node (such as \DOMElement) but almost their entire
+     * body is gone, including the `nodeType` attribute.
+     *
+     * @return bool true if node has been destroyed
+     */
+    public function isRemoved(): bool
+    {
+        return !isset($this->node->nodeType);
+    }
+
+    /**
+     * Returns the last child of node.
+     *
+     * @return SimpleHtmlDomInterface|null
+     */
+    public function lastChild()
+    {
+        /** @var \DOMNode|null $node */
+        $node = $this->node->lastChild;
+
+        if ($node === null) {
+            return null;
+        }
+
+        return new static($node);
+    }
+
+    /**
+     * Returns the next sibling of node.
+     *
+     * @return SimpleHtmlDomInterface|null
+     */
+    public function nextSibling()
+    {
+        /** @var \DOMNode|null $node */
+        $node = $this->node->nextSibling;
+
+        if ($node === null) {
+            return null;
+        }
+
+        return new static($node);
+    }
+
+    /**
+     * Returns the next sibling of node.
+     *
+     * @return SimpleHtmlDomInterface|null
+     */
+    public function nextNonWhitespaceSibling()
+    {
+        /** @var \DOMNode|null $node */
+        $node = $this->node->nextSibling;
+
+        while ($node && !\trim($node->textContent)) {
+            /** @var \DOMNode|null $node */
+            $node = $node->nextSibling;
+        }
+
+        if ($node === null) {
+            return null;
+        }
+
+        return new static($node);
+    }
+
+    /**
+     * Returns the parent of node.
+     *
+     * @return SimpleHtmlDomInterface
+     */
+    public function parentNode(): SimpleHtmlDomInterface
+    {
+        return new static($this->node->parentNode);
+    }
+
+    /**
+     * Returns the previous sibling of node.
+     *
+     * @return SimpleHtmlDomInterface|null
+     */
+    public function previousSibling()
+    {
+        /** @var \DOMNode|null $node */
+        $node = $this->node->previousSibling;
+
+        if ($node === null) {
+            return null;
+        }
+
+        return new static($node);
+    }
+
+    /**
+     * @param string|string[]|null $value <p>
+     *                                    null === get the current input value
+     *                                    text === set a new input value
+     *                                    </p>
+     *
+     * @return string|string[]|null
+     */
+    public function val($value = null)
+    {
+        if ($value === null) {
+            if (
+                $this->tag === 'input'
+                &&
+                (
+                    $this->getAttribute('type') === 'text'
+                    ||
+                    !$this->hasAttribute('type')
+                )
+            ) {
+                return $this->getAttribute('value');
+            }
+
+            if (
+                $this->hasAttribute('checked')
+                &&
+                \in_array($this->getAttribute('type'), ['checkbox', 'radio'], true)
+            ) {
+                return $this->getAttribute('value');
+            }
+
+            if ($this->node->nodeName === 'select') {
+                $valuesFromDom = [];
+                $options = $this->getElementsByTagName('option');
+                if ($options instanceof SimpleHtmlDomNode) {
+                    foreach ($options as $option) {
+                        if ($this->hasAttribute('checked')) {
+                            /** @noinspection UnnecessaryCastingInspection */
+                            $valuesFromDom[] = (string) $option->getAttribute('value');
+                        }
+                    }
+                }
+
+                if (\count($valuesFromDom) === 0) {
+                    return null;
+                }
+
+                return $valuesFromDom;
+            }
+
+            if ($this->node->nodeName === 'textarea') {
+                return $this->node->nodeValue;
+            }
+        } else {
+            /** @noinspection NestedPositiveIfStatementsInspection */
+            if (\in_array($this->getAttribute('type'), ['checkbox', 'radio'], true)) {
+                if ($value === $this->getAttribute('value')) {
+                    /** @noinspection UnusedFunctionResultInspection */
+                    $this->setAttribute('checked', 'checked');
+                } else {
+                    /** @noinspection UnusedFunctionResultInspection */
+                    $this->removeAttribute('checked');
+                }
+            } elseif ($this->node instanceof \DOMElement && $this->node->nodeName === 'select') {
+                foreach ($this->node->getElementsByTagName('option') as $option) {
+                    /** @var \DOMElement $option */
+                    if ($value === $option->getAttribute('value')) {
+                        /** @noinspection UnusedFunctionResultInspection */
+                        $option->setAttribute('selected', 'selected');
+                    } else {
+                        $option->removeAttribute('selected');
+                    }
+                }
+            } elseif ($this->node->nodeName === 'input' && \is_string($value)) {
+                // Set value for input elements
+                /** @noinspection UnusedFunctionResultInspection */
+                $this->setAttribute('value', $value);
+            } elseif ($this->node->nodeName === 'textarea' && \is_string($value)) {
+                $this->node->nodeValue = $value;
+            }
+        }
+
+        return null;
+    }
+
+    /**
+     * @param HtmlDomParser $newDocument
+     * @param bool          $removeExtraHeadTag
+     *
+     * @return HtmlDomParser
+     */
+    protected function cleanHtmlWrapper(
+        HtmlDomParser $newDocument,
+        $removeExtraHeadTag = false
+    ): HtmlDomParser {
+        if (
+            $newDocument->getIsDOMDocumentCreatedWithoutHtml()
+            ||
+            $newDocument->getIsDOMDocumentCreatedWithoutHtmlWrapper()
+        ) {
+
+            // Remove doc-type node.
+            if ($newDocument->getDocument()->doctype !== null) {
+                /** @noinspection UnusedFunctionResultInspection */
+                $newDocument->getDocument()->doctype->parentNode->removeChild($newDocument->getDocument()->doctype);
+            }
+
+            // Replace html element, preserving child nodes -> but keep the html wrapper, otherwise we got other problems ...
+            // so we replace it with "<simpleHtmlDomHtml>" and delete this at the ending.
+            $item = $newDocument->getDocument()->getElementsByTagName('html')->item(0);
+            if ($item !== null) {
+                /** @noinspection UnusedFunctionResultInspection */
+                $this->changeElementName($item, 'simpleHtmlDomHtml');
+            }
+
+            // Remove body element, preserving child nodes.
+            $body = $newDocument->getDocument()->getElementsByTagName('body')->item(0);
+            if ($body instanceof \DOMElement) {
+                $fragment = $newDocument->getDocument()->createDocumentFragment();
+
+                while ($body->childNodes->length > 0) {
+                    $tmpNode = $body->childNodes->item(0);
+                    if ($tmpNode !== null) {
+                        /** @noinspection UnusedFunctionResultInspection */
+                        $fragment->appendChild($tmpNode);
+                    }
+                }
+
+                /** @noinspection UnusedFunctionResultInspection */
+                $body->parentNode->replaceChild($fragment, $body);
+            }
+        }
+
+        // Remove head element, preserving child nodes.
+        if (
+            $removeExtraHeadTag
+            &&
+            $this->node->parentNode instanceof \DOMElement
+            &&
+            $newDocument->getIsDOMDocumentCreatedWithoutHeadWrapper()
+        ) {
+            $html = $this->node->parentNode->getElementsByTagName('head')[0] ?? null;
+
+            if ($html !== null && $this->node->parentNode->ownerDocument !== null) {
+                $fragment = $this->node->parentNode->ownerDocument->createDocumentFragment();
+
+                /** @var \DOMNode $html */
+                while ($html->childNodes->length > 0) {
+                    $tmpNode = $html->childNodes->item(0);
+                    if ($tmpNode !== null) {
+                        /** @noinspection UnusedFunctionResultInspection */
+                        $fragment->appendChild($tmpNode);
+                    }
+                }
+
+                /** @noinspection UnusedFunctionResultInspection */
+                $html->parentNode->replaceChild($fragment, $html);
+            }
+        }
+
+        return $newDocument;
+    }
+
+    /**
+     * Retrieve an external iterator.
+     *
+     * @see  http://php.net/manual/en/iteratoraggregate.getiterator.php
+     *
+     * @return SimpleHtmlDomNode
+     *                           <p>
+     *                              An instance of an object implementing <b>Iterator</b> or
+     *                              <b>Traversable</b>
+     *                           </p>
+     */
+    public function getIterator(): SimpleHtmlDomNodeInterface
+    {
+        $elements = new SimpleHtmlDomNode();
+        if ($this->node->hasChildNodes()) {
+            foreach ($this->node->childNodes as $node) {
+                $elements[] = new static($node);
+            }
+        }
+
+        return $elements;
+    }
+
+    /**
+     * Get dom node's inner html.
+     *
+     * @param bool $multiDecodeNewHtmlEntity
+     *
+     * @return string
+     */
+    public function innerXml(bool $multiDecodeNewHtmlEntity = false): string
+    {
+        return $this->getHtmlDomParser()->innerXml($multiDecodeNewHtmlEntity);
+    }
+
+    /**
+     * Normalize the given input for comparision.
+     *
+     * @param HtmlDomParser|string $input
+     *
+     * @return string
+     */
+    private function normalizeStringForComparision($input): string
+    {
+        if ($input instanceof HtmlDomParser) {
+            $string = $input->outerText();
+
+            if ($input->getIsDOMDocumentCreatedWithoutHeadWrapper()) {
+                /** @noinspection HtmlRequiredTitleElement */
+                $string = \str_replace(['<head>', '</head>'], '', $string);
+            }
+        } else {
+            $string = (string) $input;
+        }
+
+        return
+            \urlencode(
+                \urldecode(
+                    \trim(
+                        \str_replace(
+                            [
+                                ' ',
+                                "\n",
+                                "\r",
+                                '/>',
+                            ],
+                            [
+                                '',
+                                '',
+                                '',
+                                '>',
+                            ],
+                            \strtolower($string)
+                        )
                     )
                 )
-            )
-        );
-  }
-
-  /**
-   * @param HtmlDomParser $newDocument
-   *
-   * @return HtmlDomParser
-   */
-  protected function cleanHtmlWrapper(HtmlDomParser $newDocument): HtmlDomParser
-  {
-    if (
-        $newDocument->getIsDOMDocumentCreatedWithoutHtml() === true
-        ||
-        $newDocument->getIsDOMDocumentCreatedWithoutHtmlWrapper() === true
-    ) {
-
-      // Remove doc-type node.
-      if ($newDocument->getDocument()->doctype !== null) {
-        $newDocument->getDocument()->doctype->parentNode->removeChild($newDocument->getDocument()->doctype);
-      }
-
-      // Remove html element, preserving child nodes.
-      $html = $newDocument->getDocument()->getElementsByTagName('html')->item(0);
-      $fragment = $newDocument->getDocument()->createDocumentFragment();
-      if ($html !== null) {
-        while ($html->childNodes->length > 0) {
-          $fragment->appendChild($html->childNodes->item(0));
-        }
-        $html->parentNode->replaceChild($fragment, $html);
-      }
-
-      // Remove body element, preserving child nodes.
-      $body = $newDocument->getDocument()->getElementsByTagName('body')->item(0);
-      $fragment = $newDocument->getDocument()->createDocumentFragment();
-      if ($body instanceof \DOMElement) {
-        while ($body->childNodes->length > 0) {
-          $fragment->appendChild($body->childNodes->item(0));
-        }
-        $body->parentNode->replaceChild($fragment, $body);
-
-        // At this point DOMDocument still added a "<p>"-wrapper around our string,
-        // so we replace it with "<simpleHtmlDomP>" and delete this at the ending ...
-        $item = $newDocument->getDocument()->getElementsByTagName('p')->item(0);
-        if ($item !== null) {
-          $this->changeElementName($item, 'simpleHtmlDomP');
-        }
-      }
+            );
     }
-
-    return $newDocument;
-  }
-
-  /**
-   * Change the name of a tag in a "DOMNode".
-   *
-   * @param DOMNode $node
-   * @param string  $name
-   *
-   * @return DOMElement
-   */
-  protected function changeElementName(\DOMNode $node, string $name): \DOMElement
-  {
-    $newnode = $node->ownerDocument->createElement($name);
-
-    foreach ($node->childNodes as $child) {
-      $child = $node->ownerDocument->importNode($child, true);
-      $newnode->appendChild($child);
-    }
-
-    foreach ($node->attributes as $attrName => $attrNode) {
-      $newnode->setAttribute($attrName, $attrNode);
-    }
-
-    $newnode->ownerDocument->replaceChild($newnode, $node);
-
-    return $newnode;
-  }
-
-  /**
-   * Set attribute value.
-   *
-   * @param string      $name       <p>The name of the html-attribute.</p>
-   * @param string|null $value      <p>Set to NULL or empty string, to remove the attribute.</p>
-   * @param bool $strict            </p>
-   *                                $value must be NULL, to remove the attribute,
-   *                                so that you can set an empty string as attribute-value e.g. autofocus=""
-   *                                </p>
-   *
-   * @return $this
-   */
-  public function setAttribute(string $name, $value = null, bool $strict = false)
-  {
-    if (
-        ($strict === true && null === $value)
-        ||
-        ($strict === false && empty($value))
-    ) {
-      $this->node->removeAttribute($name);
-    } else {
-      $this->node->setAttribute($name, $value);
-    }
-
-    return $this;
-  }
-
-  /**
-   * Remove attribute.
-   *
-   * @param string $name <p>The name of the html-attribute.</p>
-   *
-   * @return mixed
-   */
-  public function removeAttribute(string $name)
-  {
-    $this->node->removeAttribute($name);
-
-    return $this;
-  }
-
-  /**
-   * Get dom node's plain text.
-   *
-   * @return string
-   */
-  public function text(): string
-  {
-    return $this->node->textContent;
-  }
 }
